### 2022-02-23 #4191 @editorialbot
[PRE REVIEW]: VIEWpoly: a visualization tool to integrate and explore results of polyploid genetic analysis
**Submitting author:** @cristianetaniguti (<a href="http://orcid.org/0000-0002-2021-6883">Taniguti, C. H.</a>)
**Repository:** https://github.com/mmollina/viewpoly
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
**Reviewers:**
