+++ paper/paper.bib
@@ -134,7 +134,7 @@ journal = {The Plant Genome},
 volume = {9},
 number = {2},
 pages = {plantgenome2015.08.0073},
-doi = {https://doi.org/10.3835/plantgenome2015.08.0073},
+doi = {10.3835/plantgenome2015.08.0073},
 url = {https://acsess.onlinelibrary.wiley.com/doi/abs/10.3835/plantgenome2015.08.0073},
 eprint = {https://acsess.onlinelibrary.wiley.com/doi/pdf/10.3835/plantgenome2015.08.0073},
 year = {2016}
