### 2022-02-27 #4191 @arfon
@Cristianetaniguti – as this package seems to focus heavily on visualisation I'm going to ask the JOSS editorial team to review whether this submission is in scope for us. A decision here will likely take a couple of weeks.
