### 2022-04-12 mmollina/viewpoly#23 @cpalmer718
Broken links
In a few important places, links are broken that interfere with examples/evaluating the project.

- "Available example datasets" -> "Access complete example datasets" link broken; also in vignette
- "JBrowseR" button "Open JBrowseR" fails with error "cannot open URL 'http://wenjuanwei_pc-zbz.statgen.ncsu.edu/files/genome-browser/Stuberosum_448_v4.03.gene_exons.gff3.gz'"
- "Annotation table" section fails with the same error as above, though the box is very small so the error seems to float outside the box
