### 2022-04-12 mmollina/viewpoly#24 @cpalmer718
Image creation/download
I ran into some idiosyncrasies when trying to create and download images:

- clicking download pdf on one of the images while viewing in a firefox tab opens the resultant pdf in the same browser tab, which tricked me into closing my viewpoly window when I closed it. not sure if there's a way to default it to triggering in a new tab, or going directly to download?
- from R browser window:
   - non-tiff images seem to save unconditionally in the "Documents" directory, even though the download interface shows the "Downloads" directory; this is mysterious
   - tiff images start with a suggested non-compliant image name and no file suffix; the file then vanishes, I actually have no idea where it's being saved, if it's being saved at all.
- from firefox instance: the tiff download button triggers a loading window but then does nothing, doesn't open an image or prompt for save location/name
  - checking the R session warnings, it looks like the program is trying to call "tif" as a graphics device and failing. I'm not sure if that's because I'm missing a dependency or because it's expecting "tiff"
