### 2022-04-12 mmollina/viewpoly#25 @cpalmer718
Slightly confusing error message
Presumably in the "Progeny Haplotypes" section I've selected too many inputs to find a valid candidate, but the error message "Any individual contain all the selected homolog/s" seems to be contradictory; maybe this could be reworded to clarify to the user how they might remedy the situation? In general, clear error messages are essential.
