### 2022-04-12 mmollina/viewpoly#26 @cpalmer718
Functionality documentation
Part of the submission guidelines include documentation of functions. There are roxygen headers, but they provide inconsistent coverage of parameters, usage, etc. Many of the functions are designed to never be explicitly called by a user, which is fine; but if there's any intention of expanding the feature set at some point, including this documentation is important. I'm not sure how much linting this code base has been subjected to, but some automated tools will scan and report missing chunks of docs/etc., which can really help.
