### 2022-04-12 mmollina/viewpoly#27 @cpalmer718
Paper comments
My comments here are fairly minor; I thought the document did a really good job of conveying the utility of the software.

- The Summary seems to assume a familiarity with genomics (e.g. linkage, QTLs, ploidy) that is not safe to assume in a diverse/non-specialist audience. Even a few sentences guiding people in a bit more gently would help.
- The CRAN URL is broken in the formatted document.
- In a few places, the article "the" could probably be dropped to make the prose flow a little more cleanly.
- There are very minor copy editing issues; "still impose[s]" (38-39) was the only one that drove me to write down something. In general this was really good.
