### 2022-04-12 mmollina/viewpoly#28 @cpalmer718
Many R warnings
The severity here is probably pretty low, but I got tons of R warnings in the console. Most of these can safely be ignored. However, if some of these are intended (e.g. excluding points by just restricting the axes with ggplot or something, getting the "Removed ### rows" warning), then I think suppressing the ones you're aware of in advance would be helpful, so that more critical warnings (e.g. possibly the "tif" warning) could be more noticeable.
