### 2022-04-12 mmollina/viewpoly#28 @cpalmer718
[R_console_output.txt](https://github.com/mmollina/viewpoly/files/8468575/R_console_output.txt)
