### 2022-04-12 #4242 @cpalmer718
Slightly more specific comments:

Most seriously: there are very few tests, and they really only seem to test that the app launches, not its functionality. This is obviously a challenge in GUI situations. @mikldk I think your guidance may be needed here. For a CL tool I'd say the test setup would need to be reworked from scratch, but I'm not sure that's a reasonable thing to ask here.

Assorted other comments and links:

- installation was really smooth, though I was only able to test on Windows
- broken links at various places make it hard to get example sets to use/evaluate; I've opened an issue [here](https://github.com/mmollina/viewpoly/issues/23)
- image creation and download was a bit bumpy (though when it worked it was very cool): [issue](https://github.com/mmollina/viewpoly/issues/24)
- I can't tell you how much I like the video tutorial; I wish more people would do that. I know this might be a pain, but it would be really helpful to have the screen in the video be a bit bigger; some parts were hard to see zooming by.
- very minor issue about seemingly misleading error message [here](https://github.com/mmollina/viewpoly/issues/25)
- functionality documentation is inconsistent [here](https://github.com/mmollina/viewpoly/issues/26)
- there were comments about how to submit issues and get guidance, but I didn't find any guidelines for contribution.
- fairly minor comments about paper [here](https://github.com/mmollina/viewpoly/issues/27); I really appreciate how much information the authors provide. The only possible issue is that the Summary section may not be most appropriate for a general audience; I am very familiar with QTLs and linkage and ploidy, but the generalist audience may not be
- this may just be my system, but I got a ton of warnings in the R instance I used to launch this; I'll post some in an issue [here](https://github.com/mmollina/viewpoly/issues/28)
