### 2022-04-12 1000d37
fix tif and duplicated file in wd (#24)
