+++ paper/paper.md
@@ -40,11 +40,11 @@ affiliations:
 
 # Summary
 
-Advancements in computational tools for linkage and quantitative trait loci (QTL) analysis in autopolyploids have allowed the in-depth exploration and understanding of their genetics. Despite providing advanced methods, these tools may be challenging to use when interpreting and connecting results to other sources of genomic information and when attempting to efficiently apply them in practical applications. `VIEWpoly` is an R package for visualizing, exploring, and integrating results from different polyploid computational tools using an interactive graphical user interface. The package allows users to explore linkage and QTL analysis results and integrates these with genomic information in a genome browser, facilitating the search for candidate genes. In addition, it provides features to download comprehensive tables and graphics for further analysis and presentation. `VIEWpoly` is freely available as an R package at https://CRAN.R- project.org/package=viewpoly. 
+Advancements in computational tools for linkage and quantitative trait loci (QTL) analysis in autopolyploids have allowed the in-depth exploration and understanding of their genetics. Despite providing advanced methods, these tools may be challenging to use when interpreting and connecting results to other sources of genomic information and when attempting to efficiently apply them in practical applications. `VIEWpoly` is an R package for visualizing, exploring, and integrating results from different polyploid computational tools using an interactive graphical user interface. The package allows users to explore linkage and QTL analysis results and integrates these with genomic information in a genome browser, facilitating the search for candidate genes. In addition, it provides features to download comprehensive tables and graphics for further analysis and presentation. `VIEWpoly` is freely available as an R package at https://CRAN.R-project.org/package=viewpoly. 
 
 # Statement of need
 
-In recent years, genetic analysis in autopolyploid crops has been gaining attention due to the flourishing of high-throughput genotyping technologies, which deliver massive amounts of DNA sequences. This technological advance prompted the development of many computational tools to perform an in-depth analysis of these complex genomes. Programs such as `TetraploidSNPmap` [@Hackett2017], `polymapR` [@Bourke2018], and `MAPpoly` [@Mollinari2019; @Mollinari2020] use dosage-based markers to build genetic maps and obtain the linkage phases in full-sib mapping populations. `TetraOrigin` [@Zheng2016], `PolyOrigin` [@Zheng2021], `MAPpoly` [@Mollinari2019], and `polyqtlR` [@Bourke2021] provide the probabilistic haplotype inheritance profiles of individuals in mapping populations in terms of their respective parental haplotypes. These probabilities are used to estimate the position and genetic effects of quantitative trait loci (QTL). Using this information, researchers can benefit from the inferred genome-wide genotype and phenotype relationship to make informed breeding decisions. Software programs for performing such analysis in autopolyploids include `GWASpoly` [@Rosyara2016], `QTLpoly` [@Pereira2020], `polyqtlR` [@Bourke2021], and `diaQTL` [@Amadeu2021]. Despite the increasing availability of computational and analytical tools focused on autopolyploid species, the need for programming skills to comprehensively organize, integrate, display results of these programs still impose an obstacle to exploring and using this information in practical situations. We present `VIEWpoly`, a user-friendly R package that allows easy integration, visualization, and exploration of upstream genetic and genomic analysis focused on polyploid organisms. `VIEWpoly` is a self-contained interactive R package built as `Shiny` modules using `golem` [@Fay2022] framework. It includes documentation, data examples, and comprehensive video tutorials to guide users in easy-to-follow steps to extract information from polyploid computational tools in a seamless manner. It can be deployed locally or on servers.
[…]
