+++ R/mod_genes_view.R
@@ -333,14 +333,28 @@ mod_genes_view_server <- function(input, output, session,
     } else path.wig <- NULL
     
     if(is.null(loadJBrowse_fasta()) & !is.null(loadExample())){
-        path.fa <- loadExample()$fasta
-        path.gff <- loadExample()$gff3
-        gff.dir <- tempfile()
-        download.file(loadExample()$gff3, destfile = gff.dir)
-        gff <- vroom(gff.dir, delim = "\t", skip = 3, col_names = F)
-        # Add other tracks
-        # variants_track <- track_variant()
-        # alignments_track <- track_alignments()
+      path.fa <- loadExample()$fasta
+      path.gff <- loadExample()$gff3
+      
+      ext.list <- strsplit(c(loadExample()$fasta,loadExample()$gff3), "[.]")
+      
+      ext <- sapply(ext.list, function(x) {
+        if(x[length(x)] == "gz") paste0(x[length(x)-1], ".",x[length(x)])
+      })
+      
+      fasta.dir <- paste0(tempfile(),".", ext[1])
+      download.file(loadExample()$fasta, destfile = fasta.dir)
+      download.file(paste0(loadExample()$fasta, ".fai"), destfile = paste0(fasta.dir, ".fai"))
+      path.fa <- fasta.dir
+      
+      gff.dir <- paste0(tempfile(),".", ext[2])
+      download.file(loadExample()$gff3, destfile = gff.dir)
+      path.gff <- gff.dir
+      
+      gff <- vroom(gff.dir, delim = "\t", skip = 3, col_names = F)
+      # Add other tracks
+      # variants_track <- track_variant()
+      # alignments_track <- track_alignments()
     } else if(is.null(loadJBrowse_fasta())){
       stop("Upload the genome information in upload session to access this feature.")
     }
