### 2022-04-12 mmollina/viewpoly#24 @Cristianetaniguti
Hello, thank you for all your comments.
About the first topic: It is possible to configurate the browser to avoid this behavior. I am still searching a solution in shiny to force any browser to open the pdf in a different tab.
The tiff files issue was indeed because of the typo in "tif".
The downloads were being saved duplicated in the target directory (Download) and also in the working directory (Documents). With the modifications, the copy in the working directory will be removed.
