+++ R/mod_upload.R
@@ -13,8 +13,8 @@ mod_upload_ui <- function(id){
     fluidRow(
       column(width = 12,
              div(style = "position:absolute;right:1em;", 
-                 actionButton(ns("exit"), "Exit",icon("times-circle"), class = "btn btn-danger"), br(), br(),
-                 actionButton(ns("goQTL"), "Next",icon("arrow-circle-right"), class = "btn btn-success")
+                 actionButton(ns("exit"), "Exit",icon("times-circle", verify_fa = FALSE), class = "btn btn-danger"), br(), br(),
+                 actionButton(ns("goQTL"), "Next",icon("arrow-circle-right", verify_fa = FALSE), class = "btn btn-success")
              ),
              tags$h2(tags$b("Input data")), br(),
              "Use this module to select an example dataset or to upload yours.", br(), br()
@@ -34,7 +34,7 @@ mod_upload_ui <- function(id){
              fluidPage(
                box(width = 12, solidHeader = TRUE, collapsible = TRUE, collapsed = TRUE, status="primary", title = tags$h4(tags$b("Upload linkage map files")),
                    div(style = "position:absolute;right:1em;",
-                       actionBttn(ns("reset_map"), style = "jelly", color = "royal",  size = "sm", label = "reset", icon = icon("undo-alt"))
+                       actionBttn(ns("reset_map"), style = "jelly", color = "royal",  size = "sm", label = "reset", icon = icon("undo-alt", verify_fa = FALSE))
                    ), br(), br(), 
                    box(width = 12, solidHeader = FALSE, collapsible = TRUE, collapsed = TRUE,  title = tags$h5(tags$b("Upload MAPpoly output")),
                        tags$p("Access further information about how to build a linkage maps with MAPpoly ", 
@@ -43,7 +43,7 @@ mod_upload_ui <- function(id){
                               tags$a(href= "https://cristianetaniguti.github.io/viewpoly_vignettes/VIEWpoly_tutorial.html#Upload_linkage_map_files","here")),
                        hr(),
                        div(style = "position:absolute;right:1em;",
-                           actionBttn(ns("submit_mappoly"), style = "jelly", color = "royal",  size = "sm", label = "submit MAPpoly", icon = icon("share-square")), 
+                           actionBttn(ns("submit_mappoly"), style = "jelly", color = "royal",  size = "sm", label = "submit MAPpoly", icon = icon("share-square", verify_fa = FALSE)), 
                        ), br(), br(),
                        tags$p("Object of class `mappoly.map`."), 
                        fileInput(ns("mappoly_in"), label = h6("File: my_mappoly_list.RData"), multiple = F),
@@ -55,7 +55,7 @@ mod_upload_ui <- function(id){
                               tags$a(href= "https://cristianetaniguti.github.io/viewpoly_vignettes/VIEWpoly_tutorial.html#Upload_linkage_map_files","here")),
                        hr(),
                        div(style = "position:absolute;right:1em;",
-                           actionBttn(ns("submit_polymapR"), style = "jelly", color = "royal",  size = "sm", label = "submit polymapR", icon = icon("share-square")), 
+                           actionBttn(ns("submit_polymapR"), style = "jelly", color = "royal",  size = "sm", label = "submit polymapR", icon = icon("share-square", verify_fa = FALSE)), 
                        ), br(), br(),
                        p("Indicates whether the genotype input is discrete or probabilistic."),
                        prettyRadioButtons(
@@ -82,7 +82,7 @@ mod_upload_ui <- function(id){
                    ),
                    box(width = 12, solidHeader = FALSE, collapsible = TRUE, collapsed = TRUE, title = tags$h5(tags$b("Upload linkage map files with standard format (.csv, .tsv or .tsv.gz)")),
[…]
