### 2022-04-14 c2aec0d
improving doc (#26) more stable links (#23)
