+++ R/functions_upload.R
@@ -19,8 +19,8 @@ prepare_examples <- function(example){
     })
     structure(list(map=viewmap_tetra, 
                    qtl=viewqtl_tetra,
-                   fasta = "http://gesteira.statgen.ncsu.edu/files/genome-browser/Stuberosum_448_v4.03.fa.gz",
-                   gff3 = "http://gesteira.statgen.ncsu.edu/files/genome-browser/Stuberosum_448_v4.03.gene_exons.gff3.gz"),
+                   fasta = "https://gesteira.statgen.ncsu.edu/files/genome-browser/Stuberosum_448_v4.03.fa.gz",
+                   gff3 = "https://gesteira.statgen.ncsu.edu/files/genome-browser/Stuberosum_448_v4.03.gene_exons.gff3.gz"),
               class = "viewpoly")
   }
 }
