### 2022-04-19 mmollina/viewpoly#29 @jkanche
Review comments
- Cannot access the genome browser - Jbrowser, Is there something specific about these genome files ? Can you use the one directly from UCSC or NCBI ?

- a minor UI issue, I see Next and Exit as the navigation buttons on the top right corner but not a back button. which makes me think i cannot go back but obviously not true from the menu.

- probably a user error but none of my selections on the viewQTL section of app worked

- I would prefer to see consistency in the interactions, brushing in the viewQTL is a selection but brushing on the plots in ViewMap and ViewGenome is a zoom action for the same activity.

- plotly bar overlaps with the plot borders

- is there a difference between the download (purple button) from the UI vs what plotly provides ?

- it might aesthetically make sense to show the sliders (mapRange etc) next to the dropdowns rather than a full width slider.
- another aesthetic improvement, the headers for the panels are not clear/readable, a few css changes might fix this; i did this locally, a great improvement would be to get the text and the icon on the same line.

vs now

- I wonder if the authors thought about providing access to methods for some of the visualizations used in the shiny application. I do notice these functions are documented but `noRd` & `#' @keywords internal` is probably not generating documentation for these

- can you provide an example for a viewPoly dataset ?
- do you plan on providing an instance on rsconnect for users to use the application (not everyone has the skill to install R/ViewPoly and run the application) ?
- tests seem to be very basic. Do you plan on adding tests to test various functionality ? (I am not referring to UI testing here)

reference: https://github.com/openjournals/joss-reviews/issues/4242
