### 2022-04-19 #4242 @mikldk
@cpalmer718 @Cristianetaniguti

> Most seriously: there are very few tests, and they really only seem to test that the app launches, not its functionality. This is obviously a challenge in GUI situations. @mikldk I think your guidance may be needed here. For a CL tool I'd say the test setup would need to be reworked from scratch, but I'm not sure that's a reasonable thing to ask here.

Our [guidelines](https://joss.readthedocs.io/en/latest/submitting.html#typical-paper-submission-flow) states "In particular, your software should be full-featured, well-documented, and contain procedures (such as automated tests) for checking correctness.".

So we do not have a strong requirement for automated tests, but there should be some procedures for checking correctness. One way is to supply screendumps in the help files such that the user can compare her output with those.

> ...
