### 2022-04-19 #4242 @jkanche
@mikldk @cpalmer718 I also made similar comments in my review.

It would be great to add functional testing not necessarily UI testing. UI testing is always hard for shiny applications.
