+++ paper/paper.md
@@ -40,7 +40,7 @@ affiliations:
 
 # Summary
 
-Advancements in computational tools for linkage and quantitative trait loci (QTL) analysis in autopolyploids have allowed the in-depth exploration and understanding of their genetics. Despite providing advanced methods, these tools may be challenging to use when interpreting and connecting results to other sources of genomic information and when attempting to efficiently apply them in practical applications. `VIEWpoly` is an R package for visualizing, exploring, and integrating results from different polyploid computational tools using an interactive graphical user interface. The package allows users to explore linkage and QTL analysis results and integrates these with genomic information in a genome browser, facilitating the search for candidate genes. In addition, it provides features to download comprehensive tables and graphics for further analysis and presentation. `VIEWpoly` is freely available as an R package at https://CRAN.R-project.org/package=viewpoly. 
+Autopolyploid species comprise a significant part of the agronomic market. Some examples are potato, sweet potato, roses, and blueberries. They all present more than two copies of their entire genome, which increase the complexity of genetic models. Advancements in computational tools for linkage and quantitative trait loci (QTL) analysis in autopolyploids have allowed the in-depth exploration and understanding of their genetics. Despite providing advanced methods, the results of these tools may be challenging to interpret and connect to other sources of genomic information and consequently apply them in practical breeding applications. VIEWpoly is an R package for visualizing, exploring, and integrating results from different polyploid computational tools using an interactive graphical user interface. The package allows users to explore linkage and QTL analysis results and integrates these with genomic information in a genome browser, facilitating the search for candidate genes. In addition, it provides features to download comprehensive tables and graphics for further analysis and presentation. VIEWpoly is freely available as an R package at https://CRAN.R-project.org/package=viewpoly.
 
 # Statement of need
 
@@ -52,18 +52,18 @@ In recent years, genetic analysis in autopolyploid crops has been gaining attent
 
 `VIEWpoly` comprises three main modules arranged in an intuitive interface that guides the user through the analysis \autoref{fig:flowchart}. In its current version (V 0.1.1), the package supports `polymapR` [@Bourke2018] and `MAPpoly` [@Mollinari2019; @Mollinari2020] results to display genetic maps. The QTL analysis results can be imported from `QTLpoly` [@Pereira2020], `polyqtlR` [@Bourke2021], and `diaQTL` [@Amadeu2021] packages. Maps and QTL input files can be provided in the RData format or standard text formats (CSV or TSV). `VIEWpoly` is also compatible with genome-related files supported by `JBrowseR`, including FASTA, BED, GFF3, BAM, CRAM, VCF, Wiggle, and bigWig files [@Hershberg2021]. Once the data is uploaded, the user can export the information in a single compressed file to expedite the upload process in further analysis.
 
-## `VIEWqtl`: the QTL Browser
+## `VIEWqtl`: QTL Browser
 
 The `VIEWqtl` module is designed to explore information from the QTL analysis. The first graphic contains the QTL profile, where users can display single or multiple linkage groups and phenotypes. The QTL peak positions are represented by triangles at the bottom of the graphic and can be interactively selected to explore the genetic effects of parental haplotypes. It is possible to compare pleiotropic and linked QTL effects simultaneously and select individuals carrying alleles with specific effects. In this module, users can also find a table with a summary of selected QTL characteristics, and progeny breeding values will be displayed if results from `QTLpoly` are available.
[…]
