### 2022-04-20 #4242 @mikldk
@Cristianetaniguti are you able to add functional testing as suggested by @cpalmer718 @jkanche?
