### 2022-04-20 #4242 @raivivek
@Cristianetaniguti I'm reviewing the R package right now and noticing many minor issues. I see that early reviewers have highlighted them already, so perhaps I should pull the development branch for testing to make sure I have all the latest changes. Please let me know if/when I should install the latest version in case you are in the middle of any changes. Thanks!
