### 2022-04-20 #4242 @Cristianetaniguti
Hello! Thank you all for the comments.
Yes, @raivivek , please install the main branch (devtools::install_github("https://github.com/mmollina/viewpoly"). I have already pushed some modifications according to @cpalmer718 comments. Among them are more stable links for the example files.
@mikldk I thought about  using testthat functions to check the data used in ggplot to build the graphics.
