+++ tests/testthat/test-qtls.R
@@ -1,0 +1,1 @@
+context("QTL graphics and tables")
