+++ man/check_viewpoly.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils_helpers.R
+\name{check_viewpoly}
+\alias{check_viewpoly}
+\title{Viewpoly object sanity check}
+\usage{
+check_viewpoly(viewpoly_obj)
+}
+\arguments{
+\item{viewpoly_obj}{an object of class \code{viewpoly}}
+}
+\value{
+if consistent, returns 0. If not consistent, returns a 
+        vector with a number of tests, where \code{TRUE} indicates
+        a failed test.
+}
+\description{
+Viewpoly object sanity check
+}
+\author{
+Cristiane Taniguti, \email{chtaniguti@tamu.edu}
+}
