+++ man/check_viewmap.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils_helpers.R
+\name{check_viewmap}
+\alias{check_viewmap}
+\title{Viewmap object sanity check}
+\usage{
+check_viewmap(viewmap_obj)
+}
+\arguments{
+\item{viewmap_obj}{an object of class \code{viewmap}}
+}
+\value{
+if consistent, returns 0. If not consistent, returns a 
+        vector with a number of tests, where \code{TRUE} indicates
+        a failed test.
+}
+\description{
+Viewmap object sanity check
+}
+\author{
+Cristiane Taniguti, \email{chtaniguti@tamu.edu}
+}
