+++ man/check_viewqtl.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils_helpers.R
+\name{check_viewqtl}
+\alias{check_viewqtl}
+\title{viewqtl object sanity check}
+\usage{
+check_viewqtl(viewqtl_obj)
+}
+\arguments{
+\item{viewqtl_obj}{an object of class \code{viewqtl}}
+}
+\value{
+if consistent, returns 0. If not consistent, returns a 
+        vector with a number of tests, where \code{TRUE} indicates
+        a failed test.
+}
+\description{
+viewqtl object sanity check
+}
+\author{
+Cristiane Taniguti, \email{chtaniguti@tamu.edu}
+}

[…]
