### 2022-04-20 mmollina/viewpoly#30 @Cristianetaniguti
Add tests
reference: https://github.com/openjournals/joss-reviews/issues/4242
