+++ R/functions_upload.R
@@ -71,8 +71,6 @@ prepare_map_custom_files <- function(dosages, phases, genetic_map, mks_pos=NULL)
     return(x)
   }, d.p2, d.p2.names)
   
-  any(!is.na(match(map$marker, viewpoly_obj$map$maps[[1]]$mk.names)))
-  
   if(!is.null(mks_pos)) pos <- mks_pos[,2][match(map$marker,mks_pos[,1])] else pos <- NA
   
   maps <- data.frame(mk.names = map$marker, 
@@ -83,7 +81,7 @@ prepare_map_custom_files <- function(dosages, phases, genetic_map, mks_pos=NULL)
                      ref= NA)
   
   maps <- split.data.frame(maps, maps$g.chr)
-
+  
   ploidy <- (dim(ph)[2] - 2)/2
   
   ph.p1 <- as.data.frame(select(ph, 3:(ploidy +2)))
@@ -115,15 +113,13 @@ prepare_map_custom_files <- function(dosages, phases, genetic_map, mks_pos=NULL)
 #' @keywords internal
 prepare_MAPpoly <- function(mappoly_list){
   is <- NULL
-  withProgress(message = 'Working:', value = 0, {
-    incProgress(0.1, detail = paste("Uploading mappoly data..."))
-    if(!is(mappoly_list[[1]], "mappoly.map")){
-      temp <- load(mappoly_list$datapath)
-      mappoly_list <- get(temp)
-    }
-    prep <- lapply(mappoly_list, prepare_map)
-    incProgress(0.8, detail = paste("Uploading mappoly data..."))
-  })
+  
+  if(!is(mappoly_list[[1]], "mappoly.map")){
+    temp <- load(mappoly_list$datapath)
+    mappoly_list <- get(temp)
+  }
+  prep <- lapply(mappoly_list, prepare_map)
+  
   structure(list(d.p1 = lapply(prep, "[[", 5),
                  d.p2 = lapply(prep, "[[", 6),
                  ph.p1 = lapply(prep, "[[", 3),
@@ -193,78 +189,74 @@ prepare_polymapR <- function(polymapR.dataset, polymapR.map, input.type, ploidy)
 #' @keywords internal
 prepare_QTLpoly <- function(data, remim.mod, est.effects, fitted.mod){
   is <- NULL
-  withProgress(message = 'Working:', value = 0, {
-    incProgress(0.1, detail = paste("Uploading QTLpoly data..."))
-    temp <- load(data$datapath)
-    data <- get(temp)
-    
-    temp <- load(remim.mod$datapath)
-    remim.mod <- get(temp)
-    
-    temp <- load(est.effects$datapath)
-    est.effects <- get(temp)
-    
-    temp <- load(fitted.mod$datapath)
-    fitted.mod <- get(temp)
-    
-    # Only selected markers
-    incProgress(0.3, detail = paste("Uploading QTLpoly data..."))
-    lgs.t <- lapply(data$lgs, function(x) data.frame(mk = names(x), pos = x))
-    lgs <- data.frame()
-    for(i in 1:length(lgs.t)) {
-      lgs <- rbind(lgs, cbind(LG = i,lgs.t[[i]]))
-    }
-    
-    rownames(lgs) <- NULL
-    qtl_info <- u.hat <- beta.hat <- pvalue <- profile <- effects <- data.frame()
-    for(i in 1:length(remim.mod$results)){
-      pheno = names(fitted.mod$results)[i]
-      if(!is.null(dim(fitted.mod$results[[i]]$qtls)[1])){
-        lower <- remim.mod$results[[i]]$lower[,1:2]
[…]
