+++ .github/workflows/check-standard.yml
@@ -4,10 +4,12 @@ on:
   push:
     branches:
       - main
+      - add_tests
 
   pull_request:
     branches:
       - main
+      - add_tests
 
 name: R-CMD-check
 
