+++ tests/testthat/test-genes.R
@@ -1,1 +1,1 @@
-context("Gene annotation graphic and tables")
+# Gene annotation graphic and tables
