+++ tests/testthat/test-qtls.R
@@ -1,1 +1,1 @@
-context("QTL graphics and tables")
+# QTL graphics and tables
