+++ R/functions_upload.R
@@ -254,7 +254,7 @@ prepare_QTLpoly <- function(data, remim.mod, est.effects, fitted.mod){
     profile.t <- data.frame(pheno, LOP = SIG)
     profile <- rbind(profile, profile.t)
   }
-
+  
   # Rearrange the progeny probabilities into a list
   probs <- data$Z
   
@@ -284,90 +284,82 @@ prepare_QTLpoly <- function(data, remim.mod, est.effects, fitted.mod){
 #' @keywords internal
 prepare_diaQTL <- function(scan1_list, scan1_summaries_list, fitQTL_list, BayesCI_list){
   marker <- pheno <- NULL
-  withProgress(message = 'Working:', value = 0, {
-    incProgress(0.1, detail = paste("Uploading diaQTL data..."))
+  
+  temp <- load(scan1_list$datapath)
+  scan1_list <- get(temp)
+  
+  temp <- load(scan1_summaries_list$datapath)
+  scan1_summaries_list <- get(temp)
+  
+  temp <- load(fitQTL_list$datapath)
+  fitQTL_list <- get(temp)
+  
+  temp <- load(BayesCI_list$datapath)
+  BayesCI_list <- get(temp)
+  
+  selected_mks <- scan1_list[[1]][,c(2,1,3)]
+  colnames(selected_mks) <- c("LG", "mk", "pos")
+  qtl_info <- data.frame()
+  
+  for(i in 1:length(scan1_summaries_list)){
+    temp <- cbind(pheno = names(scan1_summaries_list)[i],scan1_summaries_list[[i]]$peaks)
+    qtl_info <- rbind(qtl_info, temp)
+  }
+  
+  qtls.id <- list()
+  qtl_info2 <- data.frame()
+  
+  profile <- effects <- data.frame()
+  for(i in 1:length(fitQTL_list)){
+    qtls.id <- colnames(fitQTL_list[[i]]$effects$additive)
+    trait <- gsub("Trait: ","",fitQTL_list[[i]]$plots[[1]]$additive$labels$title)
+    qtl_temp <-  filter(qtl_info, pheno == trait & marker %in% qtls.id)
+    qtl_info2 <- rbind(qtl_info2, qtl_temp)
+    # profile
+    profile_temp <- data.frame(pheno = trait, deltaDIC = scan1_list[[which(names(scan1_list) == trait)]]$deltaDIC)
+    profile <- rbind(profile, profile_temp)
     
-    temp <- load(scan1_list$datapath)
-    scan1_list <- get(temp)
-    
-    temp <- load(scan1_summaries_list$datapath)
-    scan1_summaries_list <- get(temp)
-    
-    temp <- load(fitQTL_list$datapath)
-    fitQTL_list <- get(temp)
-    
-    temp <- load(BayesCI_list$datapath)
-    BayesCI_list <- get(temp)
-    
-    selected_mks <- scan1_list[[1]][,c(2,1,3)]
-    colnames(selected_mks) <- c("LG", "mk", "pos")
-    qtl_info <- data.frame()
-    incProgress(0.3, detail = paste("Uploading diaQTL data..."))
-    
-    for(i in 1:length(scan1_summaries_list)){
-      temp <- cbind(pheno = names(scan1_summaries_list)[i],scan1_summaries_list[[i]]$peaks)
-      qtl_info <- rbind(qtl_info, temp)
-    }
-    
-    qtls.id <- list()
-    qtl_info2 <- data.frame()
-    incProgress(0.5, detail = paste("Uploading diaQTL data..."))
-    
-    profile <- effects <- data.frame()
-    for(i in 1:length(fitQTL_list)){
[…]
