+++ R/functions_qtl.R
@@ -30,156 +30,149 @@
 plot_profile <- function(profile, qtl_info, selected_mks, pheno.col = NULL, 
                          lgs.id = NULL, by_range = TRUE, range.min = NULL, 
                          range.max = NULL, plot=TRUE, software = NULL) {
-  withProgress(message = 'Working:', value = 0, {
-    incProgress(0.1, detail = paste("building graphic..."))
-    pheno <- LG <- `Position (cM)` <- Trait <- INT <- . <- NULL
-    
-    lgs.size <- selected_mks %>% group_by(.data$LG) %>% group_map(~ tail(.x, 1)) %>% do.call(rbind, .) 
-    lgs.size <- lgs.size$pos
-    lines <- points <- thre <- map <- data.frame()
-    y.dat <- trait.names <- c()
-    count <- 0
-    
-    nphe <- length(pheno.col)
-    LGS <- selected_mks$LG
-    POS <- selected_mks$pos
-    for(p in 1:nphe) { #lines
-      TRT <- rep(unique(profile$pheno)[pheno.col[p]], length(LGS))
-      SIG <- profile[which(profile$pheno == TRT),2]
-      lines <- rbind(lines, data.frame(TRT=as.factor(TRT), LGS=LGS, POS=POS, SIG=SIG))
-    }
-    incProgress(0.3, detail = paste("building graphic..."))
-    count <- 0
-    y.dat <- c()
-    for(p in 1:nphe) { #points
-      trait.names <- unique(profile$pheno)[pheno.col[p]]
-      if(!is.null(qtl_info)) {
-        qtl_info.sub <- qtl_info %>% filter(pheno == trait.names) %>% filter(LG %in% lgs.id)
-        if(dim(qtl_info.sub)[1] > 0){
-          nqtls <-  qtl_info.sub %>% summarize(n()) 
-          TRT <- qtl_info.sub$pheno
-          LGS <- qtl_info.sub$LG
-          POS <- qtl_info.sub$Pos
-          INF <- qtl_info.sub$Pos_lower
-          SUP <- qtl_info.sub$Pos_upper
-          PVAL <- qtl_info.sub[,6]
-          H2 <- qtl_info.sub$h2
-          if(!is.null(H2)){
-            points <- rbind(points, data.frame(TRT=TRT, LGS=LGS, POS=POS, INF=INF, SUP=SUP, PVAL = PVAL, H2 = round(H2,2)))
-          } else 
-            points <- rbind(points, data.frame(TRT=TRT, LGS=LGS, POS=POS, INF=INF, SUP=SUP, PVAL = PVAL))
-          count <- count+1
-          y.dat <- c(y.dat, rep((-0.5*count), nqtls))
-        }
+  
+  pheno <- LG <- `Position (cM)` <- Trait <- INT <- . <- NULL
+  
+  lgs.size <- selected_mks %>% group_by(.data$LG) %>% group_map(~ tail(.x, 1)) %>% do.call(rbind, .) 
+  lgs.size <- lgs.size$pos
+  lines <- points <- thre <- map <- data.frame()
+  y.dat <- trait.names <- c()
+  count <- 0
+  
+  nphe <- length(pheno.col)
+  LGS <- selected_mks$LG
+  POS <- selected_mks$pos
+  for(p in 1:nphe) { #lines
+    TRT <- rep(unique(profile$pheno)[pheno.col[p]], length(LGS))
+    SIG <- profile[which(profile$pheno == TRT),2]
+    lines <- rbind(lines, data.frame(TRT=as.factor(TRT), LGS=LGS, POS=POS, SIG=SIG))
+  }
+  count <- 0
+  y.dat <- c()
+  for(p in 1:nphe) { #points
+    trait.names <- unique(profile$pheno)[pheno.col[p]]
[…]

[…]
