+++ tests/testthat/test-genes.R
@@ -1,1 +1,0 @@
-# Gene annotation graphic and tables

[…]
