+++ man/check_viewmap.Rd
@@ -20,3 +20,4 @@ Viewmap object sanity check
 \author{
 Cristiane Taniguti, \email{chtaniguti@tamu.edu}
 }
+\keyword{internal}
