+++ man/check_viewmap_values.Rd
@@ -1,13 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/utils_helpers.R
-\name{check_viewqtl}
-\alias{check_viewqtl}
-\title{viewqtl object sanity check}
+\name{check_viewmap_values}
+\alias{check_viewmap_values}
+\title{Viewpoly object sanity check}
 \usage{
-check_viewqtl(viewqtl_obj)
+check_viewmap_values(viewmap_obj, doses, phases, maps)
 }
 \arguments{
-\item{viewqtl_obj}{an object of class \code{viewqtl}}
+\item{viewpoly_obj}{an object of class \code{viewpoly}}
 }
 \value{
 if consistent, returns 0. If not consistent, returns a 
@@ -15,7 +15,7 @@ if consistent, returns 0. If not consistent, returns a
         a failed test.
 }
 \description{
-viewqtl object sanity check
+Viewpoly object sanity check
 }
 \author{
 Cristiane Taniguti, \email{chtaniguti@tamu.edu}
