+++ NAMESPACE
@@ -1,8 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-export(check_viewmap)
-export(check_viewpoly)
-export(check_viewqtl)
 export(run_app)
 import(abind)
 import(dplyr)
