### 2022-04-25 102270c
add shinytest
