+++ tests/shinytest/shinyTest_all_tabs.R
@@ -1,0 +1,21 @@
+app <- ShinyDriver$new("../../")
+app$snapshotInit("shinyTest_all_tabs")
+
+app$setInputs(viewpoly = "upload")
+app$setInputs(viewpoly = "qtl")
+app$setInputs(`qtl_view_ui_1-phenotypes` = c("PY06", "SG06", "DM06", "ST06", "NS06", "NI06", "PY07", "FM07", "SG07", "DM07", "ST07", "NS07", "NI07", "PY08", "FM08", "SG08", "DM08", "ST08", "NS08", "NI08", "PY14", "FM14", "SG14", "DM14", "ST14", "NS14", "NI14"))
+app$setInputs(`qtl_view_ui_1-group` = c("1", "5"))
+app$setInputs(`qtl_view_ui_1-effects_design` = "circle")
+app$setInputs(`qtl_view_ui_1-effects_design` = "digenic")
+app$setInputs(`qtl_view_ui_1-effects_design` = "bar")
+app$setInputs(`qtl_view_ui_1-haplo_update` = "click")
+app$setInputs(`qtl_view_ui_1-haplo` = c("Trait:PY06_LG:5_Pos:29_homolog:P1.1", "Trait:PY06_LG:5_Pos:29_homolog:P1.2"))
+app$setInputs(`qtl_view_ui_1-haplo_submit` = "click")
+app$snapshot()
+app$setInputs(viewpoly = "genes")
+app$setInputs(`genes_view_ui_1-group` = "5")
+app$snapshot()
+app$setInputs(viewpoly = "map")
+app$setInputs(`map_view_ui_1-group` = "5")
+app$setInputs(`map_view_ui_1-range` = c(20, 40))
+app$snapshot()
