+++ DESCRIPTION
@@ -67,6 +67,7 @@ LazyData: true
 RoxygenNote: 7.1.2
 Suggests: 
     testthat (>= 3.0.0),
-    vdiffr
+    vdiffr,
+    shinytest
 Config/testthat/edition: 3
 Language: en-US
