+++ tests/testthat/test-shinyTest.R
@@ -4,5 +4,6 @@ test_that("shinytest", {
   # Don't run these tests on the CRAN build servers
   skip_on_cran()
   
-  expect_pass(testApp(testnames = "shinyTest_all_tabs", compareImages = FALSE))
+  appdir <- gsub("inst", "", system.file(package = "viewpoly"))
+  expect_pass(testApp(appDir = appdir,compareImages = FALSE))
 })
