+++ inst/tests/shinytest.R
@@ -1,0 +1,3 @@
+# library(shinytest)
+# shinytest::testApp(system.file(package = "viewpoly"))
+
