+++ tests/testthat/test-shinyTest.R
@@ -4,5 +4,5 @@ test_that("shinytest", {
   # Don't run these tests on the CRAN build servers
   skip_on_cran()
   
-  expect_pass(testApp(compareImages = FALSE))
+  expect_pass(testApp(appDir = system.file(package = "viewpoly"),compareImages = FALSE))
 })
