+++ tests/testthat/test-tetra_example.R
@@ -1,219 +1,219 @@
-# test_that("upload files",{
-#   # upload examples
-#   viewpoly_obj <- prepare_examples("tetra_map")
-#   
-#   expect_equal(check_viewpoly(viewpoly_obj),0)
-#   
-#   check_viewmap_values(viewpoly_obj$map, 
-#                        c(14, 132, 139, 157, 34), 
-#                        c(36, 167, 164, 109), 
-#                        50502.07)
-#   
-#   check_viewqtl_qtlpoly_values(viewpoly_obj$qtl, 
-#                                116504, 
-#                                5.418909, 
-#                                -1.067457e-11, 
-#                                299.6155, 
-#                                0.000160791, 
-#                                2.340129e-12, 
-#                                1)
-#   
-#   # VIEWmap tests 
-#   qtl_profile_plot <- plot_profile(profile = viewpoly_obj$qtl$profile, 
-#                                    qtl_info = viewpoly_obj$qtl$qtl_info, 
-#                                    selected_mks = viewpoly_obj$qtl$selected_mks, 
-#                                    pheno.col = 2:3, 
-#                                    lgs.id = 2, 
-#                                    by_range = TRUE, 
-#                                    range.min = 30, 
-#                                    range.max = 120, 
-#                                    plot=TRUE, 
-#                                    software = NULL)
-#   
-#   expect_equal(sum(qtl_profile_plot$data$SIG, na.rm = TRUE), 84.46874, tolerance = 0.0001)
-#   
-#   maps <- lapply(viewpoly_obj$map$maps, function(x) {
-#     y <- x$l.dist
-#     names(y) <- x$mk.names
-#     y
-#   })
-#   
-#   expect_doppelganger("linkage map draw",   draw_map_shiny(left.lim = 1, 
-#                                                            right.lim = 50, 
-#                                                            ch = 1,
-#                                                            d.p1 = viewpoly_obj$map$d.p1,
-#                                                            d.p2 = viewpoly_obj$map$d.p2, 
-#                                                            maps = maps, 
-#                                                            ph.p1 = viewpoly_obj$map$ph.p1, 
-#                                                            ph.p2 = viewpoly_obj$map$ph.p2,
-#                                                            snp.names = FALSE))
-#   
-#   expect_doppelganger("linkage map draw names",   draw_map_shiny(left.lim = 1, 
-#                                                                  right.lim = 50, 
-#                                                                  ch = 1,
-#                                                                  d.p1 = viewpoly_obj$map$d.p1,
-#                                                                  d.p2 = viewpoly_obj$map$d.p2, 
-#                                                                  maps = maps, 
-#                                                                  ph.p1 = viewpoly_obj$map$ph.p1, 
-#                                                                  ph.p2 = viewpoly_obj$map$ph.p2,
-#                                                                  snp.names = TRUE))
-#   
-#   expect_doppelganger("plot map list", plot_map_list(viewpoly_obj$map))
-#   
-#   # Get max size each chromosome
-#   expect_equal(map_summary(left.lim = 1, 
-#                            right.lim = 50, 
-#                            ch = 3, 
-#                            maps = maps, 
-#                            d.p1 = viewpoly_obj$map$d.p1, 
[…]
