+++ tests/testthat/test-QTLpoly.R
@@ -1,167 +1,167 @@
-# test_that("Tests uploaded QTLpoly files",{
-# 
-#   # upload QTLpoly
-#   input.data <- remim.mod <- est.effects <- fitted.mod <- list()
-#   input.data$datapath <- tempfile()
-#   remim.mod$datapath <- tempfile()
-#   est.effects$datapath <- tempfile()
-#   fitted.mod$datapath <- tempfile()
-# 
-#   download.file("https://www.polyploids.org/sites/default/files/2022-04/tetra_QTLpoly_data.RData", destfile = input.data$datapath)
-#   download.file("https://www.polyploids.org/sites/default/files/2022-04/tetra_QTLpoly_remim.RData", destfile = remim.mod$datapath)
-#   download.file("https://www.polyploids.org/sites/default/files/2022-04/tetra_QTLpoly_effects.RData", destfile = est.effects$datapath)
-#   download.file("https://www.polyploids.org/sites/default/files/2022-04/tetra_QTLpoly_fitted.RData", destfile = fitted.mod$datapath)
-# 
-#   viewqtl_qtlpoly <- prepare_QTLpoly(data = input.data,
-#                                      remim.mod = remim.mod,
-#                                      est.effects = est.effects,
-#                                      fitted.mod = fitted.mod)
-# 
-#   expect_equal(check_viewqtl(viewqtl_qtlpoly),0)
-# 
-#   check_viewqtl_qtlpoly_values(viewqtl_qtlpoly,
-#                                116504,
-#                                5.418909,
-#                                -1.067457e-11,
-#                                299.6155,
-#                                0.000160791,
-#                                2.340129e-12,
-#                                1)
-# 
-#   #VIEWqtl tests
-#   # plotly
-#   qtl_profile_plot <- plot_profile(viewqtl_qtlpoly$profile,
-#                                    viewqtl_qtlpoly$qtl_info,
-#                                    viewqtl_qtlpoly$selected_mks,
-#                                    pheno.col = 2,
-#                                    lgs.id = 2,
-#                                    by_range = TRUE,
-#                                    range.min = 30,
-#                                    range.max = 120,
-#                                    plot=TRUE,
-#                                    software = NULL)
-# 
-#   expect_equal(sum(qtl_profile_plot$data$SIG, na.rm = TRUE), 43.81917, tolerance = 0.0001)
-# 
-#   # by range
-#   qtl_profile_data <- plot_profile(viewqtl_qtlpoly$profile,
-#                                    viewqtl_qtlpoly$qtl_info,
-#                                    viewqtl_qtlpoly$selected_mks,
-#                                    pheno.col = 2,
-#                                    lgs.id = 2,
-#                                    by_range = TRUE,
-#                                    range.min = 30,
-#                                    range.max = 120,
-#                                    plot=FALSE,
-#                                    software = NULL)
-# 
-#   expect_equal(sum(qtl_profile_data$lines$SIG, na.rm = TRUE), 43.81917, tolerance = 0.001)
-#   expect_equal(sum(qtl_profile_data$lines$`Position (cM)`), 8000.109, tolerance = 0.001)
-#   expect_equal(as.numeric(qtl_profile_data$points$PVAL), 0.000141, tolerance = 0.001)
-#   expect_equal(as.numeric(qtl_profile_data$points$H2), 0.17, tolerance = 0.001)
-#   expect_equal(as.numeric(qtl_profile_data$points$INF), 41, tolerance = 0.001)
[…]
