+++ inst/app.R
@@ -1,7 +1,2 @@
-# Launch the ShinyApp (Do not remove this comment)
-# To deploy, run: rsconnect::deployApp()
-# Or use the blue button on top of this file
-
-pkgload::load_all(export_all = FALSE,helpers = FALSE,attach_testthat = FALSE)
 options( "golem.app.prod" = TRUE)
-viewpoly::run_app() # add parameters here (if any) 
+viewpoly::run_app() 
