+++ man/check_viewqtl_qtlpoly_values.Rd
@@ -1,46 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils_helpers.R
-\name{check_viewqtl_qtlpoly_values}
-\alias{check_viewqtl_qtlpoly_values}
-\title{Check viewqtl object for QTLpoly uploaded files}
-\usage{
-check_viewqtl_qtlpoly_values(
-  viewqtl_obj,
-  pos,
-  h2,
-  u.hat,
-  beta.hat,
-  lop,
-  effect,
-  probs
-)
-}
-\arguments{
-\item{viewqtl_obj}{an object of class \code{viewqtl}}
-
-\item{pos}{sum of the position values to be matched}
-
-\item{h2}{sum of the log herdability values to be matched}
-
-\item{u.hat}{sum of the estimated u values to be matched}
-
-\item{beta.hat}{sum of the estimated beta values to be matched}
-
-\item{lop}{sum of the LOP values to be matched}
-
-\item{effect}{sum of the effects values to be matched}
-
-\item{probs}{sum of the genotype probabilties values to be matched}
-}
-\value{
-if consistent, returns 0. If not consistent, returns a 
-        vector with a number of tests, where \code{TRUE} indicates
-        a failed test.
-}
-\description{
-Check viewqtl object for QTLpoly uploaded files
-}
-\author{
-Cristiane Taniguti, \email{chtaniguti@tamu.edu}
-}
-\keyword{internal}
