+++ man/check_viewmap_values.Rd
@@ -1,23 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils_helpers.R
-\name{check_viewmap_values}
-\alias{check_viewmap_values}
-\title{Viewpoly object sanity check}
-\usage{
-check_viewmap_values(viewmap_obj, doses, phases, maps)
-}
-\arguments{
-\item{viewpoly_obj}{an object of class \code{viewpoly}}
-}
-\value{
-if consistent, returns 0. If not consistent, returns a 
-        vector with a number of tests, where \code{TRUE} indicates
-        a failed test.
-}
-\description{
-Viewpoly object sanity check
-}
-\author{
-Cristiane Taniguti, \email{chtaniguti@tamu.edu}
-}
-\keyword{internal}
