+++ DESCRIPTION
@@ -69,7 +69,6 @@ RoxygenNote: 7.1.2
 Suggests: 
     testthat (>= 3.0.0),
     vdiffr,
-    shinytest,
-    devtools
+    shinytest
 Config/testthat/edition: 3
 Language: en-US

[…]
