### 2022-04-26 d0017e4
reduce snapshots
