+++ inst/tests/shinytest/all_tabs.R
@@ -14,9 +14,6 @@ app$setInputs(`qtl_view_ui_1-effects_design` = "bar")
 app$setInputs(`qtl_view_ui_1-haplo_update` = "click")
 app$setInputs(`qtl_view_ui_1-haplo` = c("Trait:PY06_LG:5_Pos:29_homolog:P1.1", "Trait:PY06_LG:5_Pos:29_homolog:P1.2"))
 app$setInputs(`qtl_view_ui_1-haplo_submit` = "click")
-app$snapshot()
 app$setInputs(viewpoly = "genes")
-app$snapshot()
 app$setInputs(viewpoly = "map")
 app$setInputs(`map_view_ui_1-group` = "5")
-app$snapshot()
