+++ R/mod_qtl_view.R
@@ -196,7 +196,7 @@ mod_qtl_view_server <- function(input, output, session,
                            by_range=F, plot = F)
       })
     } else
-      stop("Upload the QTL information in upload session to access this feature.")
+      stop(safeError("Upload the QTL information in upload session to access this feature."))
   })
   
   output$plot_qtl <- renderPlot({
@@ -211,7 +211,7 @@ mod_qtl_view_server <- function(input, output, session,
       if(!is.null(input$plot_brush)){
         df <- brushedPoints(qtl.data()[[2]], input$plot_brush, xvar = "x", yvar = "y.dat")
       } else {
-        stop("Select a point or region on QTL profile graphic.") 
+        stop(safeError("Select a point or region on QTL profile graphic."))
       }
       withProgress(message = 'Working:', value = 0, {
         incProgress(0.5, detail = paste("Getting data..."))
@@ -225,7 +225,7 @@ mod_qtl_view_server <- function(input, output, session,
                              design = input$effects_design)
       })
     } else 
-      stop("Upload the QTL information in upload session to access this feature.")
+      stop(safeError("Upload the QTL information in upload session to access this feature."))
   })
   
   output$effects <- renderPlot({
@@ -240,7 +240,7 @@ mod_qtl_view_server <- function(input, output, session,
       if(!is.null(input$plot_brush)){
         dframe <- brushedPoints(qtl.data()[[2]], input$plot_brush, xvar = "x", yvar = "y.dat")
       } else {
-        stop("Select a point or region on QTL profile graphic.")
+        stop(safeError("Select a point or region on QTL profile graphic."))
       }
       counts <- nrow(dframe)
       counts <- ceiling(counts/4)
@@ -257,7 +257,7 @@ mod_qtl_view_server <- function(input, output, session,
       }
       size
     } else 
-      stop("Upload the QTL information in upload session to access this feature.")
+      stop(safeError("Upload the QTL information in upload session to access this feature."))
   })
   
   output$plot.ui <- renderUI({
@@ -315,10 +315,10 @@ mod_qtl_view_server <- function(input, output, session,
   })
   
   haplo_data <- eventReactive(input$haplo_submit, {
-    if(all(input$haplo == paste0("Feature not implemented for software: ", loadQTL()$software))) stop(paste0("Feature not implemented for software: ", loadQTL()$software))
-    if(all(input$haplo == "Click on `update available haplotype` to update")) stop("Click on `update available haplotype` to update")
-    if(all(input$haplo == "Select QTL in the profile graphic to update")) stop("Select QTL in the profile graphic to update")
-    if(all(input$haplo == "Select `bar` design to access this feature.")) stop("Select `bar` design to access this feature.")
+    if(all(input$haplo == paste0("Feature not implemented for software: ", loadQTL()$software))) stop(safeError(paste0("Feature not implemented for software: ", loadQTL()$software)))
+    if(all(input$haplo == "Click on `update available haplotype` to update")) stop(safeError("Click on `update available haplotype` to update"))
+    if(all(input$haplo == "Select QTL in the profile graphic to update")) stop(safeError("Select QTL in the profile graphic to update"))
+    if(all(input$haplo == "Select `bar` design to access this feature.")) stop(safeError("Select `bar` design to access this feature."))
     p <- select_haplo(input$haplo, loadQTL()$probs, loadQTL()$selected_mks, effects.data())
     counts <- ceiling(length(p)/3)
     if(counts == 0) counts <- 1
@@ -345,7 +345,7 @@ mod_qtl_view_server <- function(input, output, session,
       if(!is.null(input$plot_brush)){
[…]
