+++ inst/tests/shinytest.R
@@ -1,3 +1,3 @@
 library(shinytest)
-shinytest::testApp("../")
+shinytest::testApp("../", compareImages = FALSE)
 
