+++ inst/tests/shinytest/mytest.R
@@ -1,0 +1,38 @@
+app <- ShinyDriver$new("../../")
+app$snapshotInit("mytest")
+
+app$setInputs(viewpoly = "upload")
+app$setInputs(viewpoly = "qtl")
+app$snapshot(list(output = "qtl_view_ui_1-plot_qtl"))
+app$setInputs(`qtl_view_ui_1-phenotypes` = c("PY06", "SG06", "DM06", "ST06", "NS06", "NI06", "PY07", "FM07", "SG07", "DM07", "ST07", "NS07", "NI07", "PY08", "FM08", "SG08", "DM08", "ST08", "NS08", "NI08", "PY14", "FM14", "SG14", "DM14", "ST14", "NS14", "NI14"))
+app$snapshot(list(output = "qtl_view_ui_1-plot_qtl"))
+app$setInputs(`qtl_view_ui_1-group` = c("1", "5"))
+app$snapshot(list(output = "qtl_view_ui_1-plot_qtl"))
+app$snapshot(list(output = "qtl_view_ui_1-effects"))
+app$snapshot(list(output = "qtl_view_ui_1-effects"))
+app$setInputs(`qtl_view_ui_1-effects_design` = "circle")
+app$snapshot(list(output = "qtl_view_ui_1-effects"))
+app$snapshot(list(output = "qtl_view_ui_1-effects"))
+app$setInputs(`qtl_view_ui_1-effects_design` = "digenic")
+app$snapshot(list(output = "qtl_view_ui_1-effects"))
+app$snapshot(list(output = "qtl_view_ui_1-effects"))
+app$setInputs(`qtl_view_ui_1-haplo_update` = "click")
+app$setInputs(`qtl_view_ui_1-effects_design` = "bar")
+app$setInputs(`qtl_view_ui_1-haplo_update` = "click")
+app$setInputs(`qtl_view_ui_1-haplo_submit` = "click")
+app$setInputs(`qtl_view_ui_1-haplo` = c("Trait:PY06_LG:5_Pos:29_homolog:P1.1", "Trait:PY06_LG:5_Pos:29_homolog:P1.2"))
+app$setInputs(`qtl_view_ui_1-haplo_submit` = "click")
+app$snapshot(list(output = "qtl_view_ui_1-haplotypes"))
+app$snapshot(list(output = "qtl_view_ui_1-haplotypes"))
+app$snapshot(list(output = "qtl_view_ui_1-breeding_values"))
+app$snapshot(list(output = "qtl_view_ui_1-info"))
+app$setInputs(viewpoly = "genes")
+app$setInputs(`genes_view_ui_1-phenotypes` = c("PY06", "SG06", "DM06", "ST06", "NS06", "NI06", "PY07", "FM07", "SG07", "DM07", "ST07", "NS07", "NI07", "PY08", "FM08", "SG08", "DM08", "ST08", "NS08", "NI08", "PY14", "FM14", "SG14", "DM14", "ST14", "NS14", "NI14"))
+app$snapshot(list(output = "genes_view_ui_1-plot_qtl"))
+app$snapshot(list(output = "genes_view_ui_1-plot_pos"))
+app$setInputs(viewpoly = "map")
+app$snapshot(list(output = "map_view_ui_1-plot_qtl"))
+app$snapshot(list(output = "map_view_ui_1-plot_map"))
+app$snapshot(list(output = "map_view_ui_1-parents_haplo"))
+app$snapshot(list(output = "map_view_ui_1-summary"))
+app$snapshot(list(output = "map_view_ui_1-map_summary"))
