+++ inst/tests/shinytest/mytest.R
@@ -32,7 +32,6 @@ app$snapshot(list(output = "genes_view_ui_1-plot_qtl"))
 app$snapshot(list(output = "genes_view_ui_1-plot_pos"))
 app$setInputs(viewpoly = "map")
 app$snapshot(list(output = "map_view_ui_1-plot_qtl"))
-app$snapshot(list(output = "map_view_ui_1-plot_map"))
 app$snapshot(list(output = "map_view_ui_1-parents_haplo"))
 app$snapshot(list(output = "map_view_ui_1-summary"))
 app$snapshot(list(output = "map_view_ui_1-map_summary"))
