+++ inst/tests/shinytest/mytest.R
@@ -7,7 +7,6 @@ app$snapshot(list(output = "qtl_view_ui_1-plot_qtl"))
 app$setInputs(`qtl_view_ui_1-phenotypes` = c("PY06", "SG06", "DM06", "ST06", "NS06", "NI06", "PY07", "FM07", "SG07", "DM07", "ST07", "NS07", "NI07", "PY08", "FM08", "SG08", "DM08", "ST08", "NS08", "NI08", "PY14", "FM14", "SG14", "DM14", "ST14", "NS14", "NI14"))
 app$snapshot(list(output = "qtl_view_ui_1-plot_qtl"))
 app$setInputs(`qtl_view_ui_1-group` = c("1", "5"))
-app$snapshot(list(output = "qtl_view_ui_1-plot_qtl"))
 app$snapshot(list(output = "qtl_view_ui_1-effects"))
 app$snapshot(list(output = "qtl_view_ui_1-effects"))
 app$setInputs(`qtl_view_ui_1-effects_design` = "circle")
