+++ R/mod_qtl_view.R
@@ -18,8 +18,11 @@ mod_qtl_view_ui <- function(id){
         fluidRow(
           column(width = 12,
                  div(style = "position:absolute;right:1em;", 
-                     actionButton(ns("exit"), "Exit",icon("times-circle", verify_fa = FALSE), class = "btn btn-danger"), br(), br(),
-                     actionButton(ns("goGenes"), "Next",icon("arrow-circle-right", verify_fa = FALSE), class = "btn btn-success")
+                     div(style = "position:absolute;right:1em;",
+                         actionButton(ns("exit"), "Exit",icon("times-circle", verify_fa = FALSE), class = "btn btn-danger")), br(), br(), br(),
+                     div(
+                       actionButton(ns("goUploads"), "Previous",icon("arrow-circle-left", verify_fa = FALSE), class = "btn btn-success"),
+                       actionButton(ns("goGenes"), "Next",icon("arrow-circle-right", verify_fa = FALSE), class = "btn btn-success"))
                  )
           ),
           tags$h2(tags$b("VIEWqtl")), br(), hr(),
@@ -180,6 +183,11 @@ mod_qtl_view_server <- function(input, output, session,
   observeEvent(input$goGenes, {
     updateTabsetPanel(session = parent_session, inputId = "viewpoly",
                       selected = "genes")
+  })
+  
+  observeEvent(input$goUploads, {
+    updateTabsetPanel(session = parent_session, inputId = "viewpoly",
+                      selected = "upload")
   })
   
   qtl.data <- reactive({
