+++ R/mod_genes_view.R
@@ -34,7 +34,7 @@ mod_genes_view_ui <- function(id){
           
           column(6,
                  column(6,
-                        box(width = 12, solidHeader = TRUE, status="info", title = h4("Select phenotypes"),
+                        box(width = 12, solidHeader = TRUE, status="info", title = "Select phenotypes",
                             pickerInput(ns("phenotypes"),
                                         label = h4("Phenotypes:"),
                                         choices = "This will be updated",
@@ -48,7 +48,7 @@ mod_genes_view_ui <- function(id){
                         ), br(),
                  ),
                  column(6,
-                        box(width = 12, solidHeader = TRUE, status="info", title = h4("Select linkage group"),
+                        box(width = 12, solidHeader = TRUE, status="info", title = "Select linkage group",
                             selectInput(inputId = ns("group"), label = p("Linkage group:"), choices = 1:15, selected = 1),
                         ), br(),
                  )
@@ -59,7 +59,7 @@ mod_genes_view_ui <- function(id){
                       value = c(0, 20), step = 1), 
           uiOutput(ns("interval"))
         ),
-        box(width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = TRUE, status="primary", title = h4("QTL profile"),
+        box(width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = TRUE, status="primary", title = "QTL profile",
             column(2,
                    downloadBttn(ns('bn_download'), style = "gradient", color = "royal")
             ),
@@ -71,7 +71,7 @@ mod_genes_view_ui <- function(id){
                    plotlyOutput(ns("plot_qtl"))
             )
         ), br(),
-        box(width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = FALSE, status="primary", title = h4("Linkage Map position (cM) x Physical position (Mp)"),
+        box(width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = FALSE, status="primary", title = "Linkage Map position (cM) x Physical position (Mp)",
             column(2,
                    downloadBttn(ns('bn_download_phi'), style = "gradient", color = "royal")
             ),
@@ -83,7 +83,7 @@ mod_genes_view_ui <- function(id){
                    plotlyOutput(ns("plot_pos"))
             )
         ), br(),
-        box(width = 12, height = 1000, solidHeader = TRUE, collapsible = TRUE,  collapsed = FALSE, status="primary", title = h4("JBrowseR"),
+        box(width = 12, height = 1000, solidHeader = TRUE, collapsible = TRUE,  collapsed = FALSE, status="primary", title = "JBrowseR",
             column(12,
                    column(6,
                           actionButton(ns("create_server"), "Open JBrowseR",icon("power-off", verify_fa = FALSE))
@@ -98,7 +98,7 @@ mod_genes_view_ui <- function(id){
             column(12, br(), hr(),
                    JBrowseROutput(ns("browserOutput"))
             ), br()),
-        box(width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = FALSE, status="primary", title = h4("Annotation table"),
+        box(width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = FALSE, status="primary", title = "Annotation table",
             DT::dataTableOutput(ns("genes_ano"))
         )
       )
