+++ R/mod_upload.R
@@ -16,15 +16,15 @@ mod_upload_ui <- function(id){
                  div(style = "position:absolute;right:1em;",
                      actionButton(ns("exit"), "Exit",icon("times-circle", verify_fa = FALSE), class = "btn btn-danger")), br(), br(), br(),
                  div(
-                     actionButton(ns("goAbout"), "Previous",icon("arrow-circle-left", verify_fa = FALSE), class = "btn btn-success"), 
-                     actionButton(ns("goQTL"), "Next",icon("arrow-circle-right", verify_fa = FALSE), class = "btn btn-success"))
+                     actionButton(ns("goAbout"), "Previous",icon("arrow-circle-left", verify_fa = FALSE), class = "btn btn-primary"), 
+                     actionButton(ns("goQTL"), "Next",icon("arrow-circle-right", verify_fa = FALSE), class = "btn btn-primary"))
              ),
              tags$h2(tags$b("Input data")), br(),
              "Use this module to select an example dataset or to upload yours.", br(), br()
       ), br(),
       column(width = 12,
              fluidPage(
-               box(width = 12, solidHeader = TRUE, collapsible = TRUE, collapsed = FALSE, status="primary", title = tags$h4(tags$b("Available example datasets")),
+               box(width = 12, solidHeader = TRUE, collapsible = TRUE, collapsed = FALSE, status="primary", title = tags$b("Available example datasets"),
                    radioButtons(ns("example_map"), label = p("They contain the entire linkage map and QTL analysis but just a subset of individuals."), 
                                 choices = c("Potato - Atlantic x B1829-5" = "tetra_map"),
                                 selected = "tetra_map"), br(), br(), hr(),
@@ -35,11 +35,11 @@ mod_upload_ui <- function(id){
       ), br(),
       column(width = 12,
              fluidPage(
-               box(width = 12, solidHeader = TRUE, collapsible = TRUE, collapsed = TRUE, status="primary", title = tags$h4(tags$b("Upload linkage map files")),
+               box(width = 12, solidHeader = TRUE, collapsible = TRUE, collapsed = TRUE, status="primary", title = tags$b("Upload linkage map files"),
                    div(style = "position:absolute;right:1em;",
                        actionBttn(ns("reset_map"), style = "jelly", color = "royal",  size = "sm", label = "reset", icon = icon("undo-alt", verify_fa = FALSE))
                    ), br(), br(), 
-                   box(width = 12, solidHeader = FALSE, collapsible = TRUE, collapsed = TRUE,  title = tags$h5(tags$b("Upload MAPpoly output")),
+                   box(width = 12, solidHeader = FALSE, collapsible = TRUE, collapsed = TRUE, status="primary",  title = tags$b("Upload MAPpoly output"),
                        tags$p("Access further information about how to build a linkage maps with MAPpoly ", 
                               tags$a(href= "https://rpubs.com/mmollin/tetra_mappoly_vignette","here")), br(),
                        tags$p("Access a example code of how to obtain these inputs using MAPpoly functions ", 
@@ -51,7 +51,7 @@ mod_upload_ui <- function(id){
                        tags$p("Object of class `mappoly.map`."), 
                        fileInput(ns("mappoly_in"), label = h6("File: my_mappoly_list.RData"), multiple = F),
                    ),
-                   box(width = 12, solidHeader = FALSE, collapsible = TRUE, collapsed = TRUE,  title = tags$h5(tags$b("Upload polymapR output")),
+                   box(width = 12, solidHeader = FALSE, collapsible = TRUE, collapsed = TRUE,  status="primary",title = tags$b("Upload polymapR output"),
[…]

[…]
