+++ R/mod_qtl_view.R
@@ -21,8 +21,8 @@ mod_qtl_view_ui <- function(id){
                      div(style = "position:absolute;right:1em;",
                          actionButton(ns("exit"), "Exit",icon("times-circle", verify_fa = FALSE), class = "btn btn-danger")), br(), br(), br(),
                      div(
-                       actionButton(ns("goUploads"), "Previous",icon("arrow-circle-left", verify_fa = FALSE), class = "btn btn-success"),
-                       actionButton(ns("goGenes"), "Next",icon("arrow-circle-right", verify_fa = FALSE), class = "btn btn-success"))
+                       actionButton(ns("goUploads"), "Previous",icon("arrow-circle-left", verify_fa = FALSE), class = "btn btn-primary"),
+                       actionButton(ns("goGenes"), label = div("Next", icon("arrow-circle-right", verify_fa = FALSE)), class = "btn btn-primary"))
                  )
           ),
           tags$h2(tags$b("VIEWqtl")), br(), hr(),
