+++ .github/workflows/check-standard.yml
@@ -34,6 +34,12 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
 
     steps:
+      # do not convert line feeds in windows
+      - name: Windows git setup
+        if: runner.os == 'Windows'
+        run:
+          git config --global core.autocrlf false
+          
       - uses: actions/checkout@v2
 
       - uses: r-lib/actions/setup-r@v1
