+++ .github/workflows/check-standard.yml
@@ -76,6 +76,7 @@ jobs:
           remotes::install_cran("covr")
           remotes::install_cran("webshot")
           webshot::install_phantomjs()
+          webdriver::run_phantomjs(timeout = 10000)
         shell: Rscript {0}
 
       - name: Check
