### 2022-04-28 mmollina/viewpoly#30 @Cristianetaniguti
Now, it contains testthat tests for the methods used to generate the graphics (the functions_* scripts in R/).

For some of the graphics checking just the input dataset was not efficient, then I used vdiffr package to convert the images to SVG and then match expected and current.

The most difficult code to test was the reactive server functions. I used the shinytest package to create snapshots of the app to be compared. However, it seems to be very sensitive to the machine it runs. It runs all tests perfectly on my computer but using GitHub actions, it usually captures small differences in image sizes. To overcome this, I created two shinytest files, one that runs the test just for the basic interface and doesn`t break the Github actions (github_actions_tests.R), and another to be run just locally (local_tests.R). I am still trying to find a way to set a tolerance or another way to make it work in all systems.
