### 2022-04-28 mmollina/viewpoly#29 @Cristianetaniguti
Hello,

First, thank you for your review!

* Cannot access the genome browser - Jbrowser, Is there something specific about these genome files ? Can you use the one directly from UCSC or NCBI ?
Nothing special about this genome, it is stored in the Phytozome database. It makes available a JGI data portal API to users to download the data. Because there is an API behind it, direct links cannot be generated. Also, JBrowseR requires some index files in addition to the genome files. We just download the genome via web platform, created the index files, and stored them on our server.

*a minor UI issue, I see Next and Exit as the navigation buttons on the top right corner but not a back button. which makes me think i cannot go back but obviously not true from the menu.
Changes made in 099d2f5b52ae02b020d6307e37e01aa62d6db57d

* probably a user error but none of my selections on the viewQTL section of app worked
The selection only works if there is/are QTLs detected. I changed the error message to clarify: 6be326937d6b867aba15176528ee56b44098d6e4

* I would prefer to see consistency in the interactions, brushing in the viewQTL is a selection but brushing on the plots in ViewMap and ViewGenome is a zoom action for the same activity.
I changed the default action for the ViewMap and ViewGenome to avoid brushing. I think the main purpose of this graphic is to guide the range setting and showing the exact significance value at each position (hover) is the most useful action. The zoom in and out can still be accessed using the Plotly mode bar. Commit: 0ddb08dfe8894749a572f2e3b08a63fe56ba8962

* plotly bar overlaps with the plot borders
Fixed: 0ddb08dfe8894749a572f2e3b08a63fe56ba8962

* is there a difference between the download (purple button) from the UI vs what plotly provides ?
No difference, it was redundant, I removed the download option off the plotly mode bar: 0ddb08dfe8894749a572f2e3b08a63fe56ba8962

* it might aesthetically make sense to show the sliders (mapRange etc) next to the dropdowns rather than a full width slider.
We kept the mapRange at full size to try to create collinearity with the following graphics x axis. The range selected is colored in the QTL profile and cM x Mb and selected in the JBrowse viewer.  The only graphic that this doesn´t happen is in the map section, that the graphic shows a zoom with only the range selected and not a highlight of the region.

* another aesthetic improvement, the headers for the panels are not clear/readable, a few css changes might fix this; I did this locally, a great improvement would be to get the text and the icon on the same line.
Fixed: e09a217de56e40f529d3c578c61edfcf5cd821a0

* I wonder if the authors thought about providing access to methods for some of the visualizations used in the shiny application. I do notice these functions are documented but noRd & #' @keywords internal is probably not generating documentation for these
We improved the documentation for all functions and removed the noRd to solve issue #26. Now, you can find documents for all functions in the man/ directory.

* can you provide an example for a viewPoly dataset?
We fixed the link to the example files: https://www.polyploids.org/input-tests
Sorry that I didn`t realize before that it was broken. Now, it should be more stable.

* do you plan on providing an instance on rsconnect for users to use the application (not everyone has the skill to install R/ViewPoly and run the application)?
Thanks for this idea! Now, it is also available here: https://cris-taniguti.shinyapps.io/viewpoly/
I updated the README to refer to it.

* tests seem to be very basic. Do you plan on adding tests to test various functionality? (I am not referring to UI testing here)
Please, check #30
