+++ tests/testthat/test-shinyTest.R
@@ -4,5 +4,9 @@ test_that("shinytest", {
   # Don't run these tests on the CRAN build servers
   skip_on_cran()
   
-  expect_pass(testApp(appDir = system.file(package = "viewpoly"), testnames = c("github_actions_tests"),compareImages = FALSE))
+  if(!grepl("inst", system.file(package = "viewpoly"))) 
+    appdir <- paste0(system.file(package = "viewpoly"), "/inst")
+  else appdir <-  system.file(package = "viewpoly")
+  
+  expect_pass(testApp(appDir = appdir, testnames = c("github_actions_tests"),compareImages = FALSE))
 })
