### 2022-04-28 23886d1
rm template functions
