+++ R/golem_utils_ui.R
@@ -1,376 +1,0 @@
-#' Turn an R list into an HTML list
-#'
-#' @param list An R list
-#' @param class a class for the list
-#' 
-#' @return an HTML list
-#' @noRd
-#' 
-#' @examples
-#' list_to_li(c("a","b"))
-#'
-#' @importFrom shiny tags tagAppendAttributes tagList
-list_to_li <- function(list, class = NULL){
-  if (is.null(class)){
-    tagList(
-      lapply(
-        list, 
-        tags$li
-      )
-    )
-  } else {
-    res <- lapply(
-      list, 
-      tags$li
-    )
-    res <- lapply(
-      res, 
-      function(x) {
-        tagAppendAttributes(
-          x, 
-          class = class
-        )
-      }
-    )
-    tagList(res)
-  }
-  
-}
-#' Turn an R list into corresponding HTML paragraph tags
-#'
-#' @param list an R list
-#' @param class a class for the paragraph tags
-#' 
-#' @return An HTML tag
-#' @noRd
-#' 
-#' @examples 
-#' list_to_p(c("This is the first paragraph", "this is the second paragraph"))
-#' 
-#' @importFrom shiny tags tagAppendAttributes tagList
-#' 
-list_to_p <- function(list, class = NULL){
-  if (is.null(class)){
-    tagList(
-      lapply(
-        list, 
-        tags$p
-      )
-    )
-  } else {
-    res <- lapply(
-      list, 
-      tags$p
-    )
-    res <- lapply(
-      res, 
-      function(x) { 
-        tagAppendAttributes(
-          x, 
-          class = class
-        )
-      }
-    )
-    tagList(res)
-  }
-  
-}
-
-#' @importFrom shiny tags tagAppendAttributes tagList
-named_to_li <- function(list, class = NULL){
-  if(is.null(class)){
-    res <- mapply(
-      function(x, y){
-        tags$li(
-          HTML(
-            sprintf("<b>%s:</b> %s", y, x)
-          )
-        )
-      },
-      list, 
-      names(list), 
-      SIMPLIFY = FALSE
-    )
-    tagList(res)
-  } else {
-    res <- mapply(
-      function(x, y){
-        tags$li(
-          HTML(
-            sprintf("<b>%s:</b> %s", y, x)
-          )
-        )
-      },
-      list,
-      names(list), 
-      SIMPLIFY = FALSE
-    )
-    res <- lapply(
-      res, 
-      function(x) {
-        tagAppendAttributes(
-          x, 
-          class = class
-        )
-      }
-    )
-    tagList(res)
-  }
-}
-
-#' Remove a tag attribute
-#'
-#' @param tag the tag
-#' @param ... the attributes to remove
-#'
-#' @return a new tag
-#' @noRd
-#' 
-#' @examples
-#' a <- shiny::tags$p(src = "plop", "pouet")
-#' tagRemoveAttributes(a, "src")
-tagRemoveAttributes <- function(tag, ...) {
-  attrs <- as.character(list(...))
-  for (i in seq_along(attrs)) {
-    tag$attribs[[ attrs[i] ]] <- NULL
-  }
-  tag
-}
-
-#' Hide or display a tag
-#' 
-#' @param tag the tag
-#' 
-#' @return a tag
-#' @noRd
-#' 
-#' @examples
-#' ## Hide
-#' a <- shiny::tags$p(src = "plop", "pouet")
-#' undisplay(a)
-#' b <- shiny::actionButton("go_filter", "go")
-#' undisplay(b)
-#' 
-#' @importFrom shiny tagList
-undisplay <- function(tag) {
-  # if not already hidden
-  if (
-    !is.null(tag$attribs$style) && 
[…]
