+++ inst/app.R
@@ -1,2 +1,3 @@
 options( "golem.app.prod" = TRUE)
 viewpoly::run_app() 
+
