### 2022-05-19 #4242 @Cristianetaniguti
Hello, I just push a simple Contributing Guidelines (mmollina/viewpoly@18ed13912d75f75882c8c9c6358a6f09954866f8) and with this, I think I cover all comments and suggestions made by @jkanche and @cpalmer718.
