### 2022-05-20 mmollina/viewpoly#34 @raivivek
Review #3 comments
Hi @Cristianetaniguti! First of all, I want to say that nice work on your JOSS submission. I apologize for incredibly delayed response, but the app has improved a lot since I saw it so I only have a few minor / nice-to-have feedback. Let me know your thoughts!

**main comments**
- With the example dataset, the QTL tab is not working properly. For the "Effects", "Breeding values", and "QTL summary" modules I see an error: _Select at least one triangle on the bottom of the QTL profile graphic. The triangles refer to QTL peaks detected. You can click and brush your cursor to select more than one._
- It'd be nice to add a prompt to Exit button before really quitting. I was confused at one point if it'll to return to home / reset session and not quit the entire app. It will also be helpful to move the button to header since it controls the entire app not just a tab.
- Increase the DPI of downloaded figure. They are large and as such when reduced to a comfortable size, the text is hard to read.
- Next/Previous should clarify that it will go to the new tab and not to the modules within the tab. Could say “Go to QTL tab” or “Go back to Input data” etc
- In each tab, it will be helpful to indicate required input boxes, if any using asterisks

**minor/nice to have**

- Update collapse/expand behavior of module tab to also work with the module header text and not just the +/- sign; it should make the app a lot more functional in my opinion
- It is also not clear which modules are expanded or collapsed when I switch/open a tab. For example, in the _Map_ tab, "QTL profile" is collapsed but second module is expanded.
- Improve ShinyApp exception handling; should you encounter an error, capture the exception and show a message/prompt on front-end that something happened and one should refresh/restart; along with the error message if possible so that users can report issues to you
- It'd be nice to have additional input validation. For example, if someone clicks submit VIEWPoly without actually adding data, the instance should not crash. Would recommend checking for file existence in all places where you process data.
