### 2022-05-20 #4242 @raivivek
Hi @Cristianetaniguti! I added a [few comments](https://github.com/mmollina/viewpoly/issues/34) to your repository based on my review! Let me know if you have any questions, but other than that, it's looks be in very good shape.
