+++ NAMESPACE
@@ -40,13 +40,9 @@ importFrom(plotly,renderPlotly)
 importFrom(reshape2,acast)
 importFrom(reshape2,dcast)
 importFrom(reshape2,melt)
-importFrom(shiny,HTML)
 importFrom(shiny,NS)
-importFrom(shiny,column)
 importFrom(shiny,shinyApp)
-importFrom(shiny,tagAppendAttributes)
 importFrom(shiny,tagList)
-importFrom(shiny,tags)
 importFrom(shinyjs,inlineCSS)
 importFrom(stats,chisq.test)
 importFrom(stats,dhyper)
