### 2022-05-24 6ed54ed
add required inputs info #34
