+++ R/mod_qtl_view.R
@@ -26,7 +26,7 @@ mod_qtl_view_ui <- function(id){
           tags$h2(tags$b("VIEWqtl")), br(), hr(),
           column(6,
                  column(6,
-                        box(width = 12, solidHeader = TRUE, status="info", title = "Select linkage group/s",
+                        box(width = 12, solidHeader = TRUE, status="info", title = "Select linkage group/s*",
                             pickerInput(ns("group"),
                                         label = h6("Linkage group/s:"),
                                         choices = "This will be updated",
@@ -40,7 +40,7 @@ mod_qtl_view_ui <- function(id){
                         )
                  ),
                  column(6,
-                        box(width = 12, solidHeader = TRUE, status="info", title = "Select phenotype/s",
+                        box(width = 12, solidHeader = TRUE, status="info", title = "Select phenotype/s*",
                             pickerInput(ns("phenotypes"),
                                         label = h6("Phenotype/s:"),
                                         choices = "This will be updated",
@@ -58,7 +58,7 @@ mod_qtl_view_ui <- function(id){
                  box(width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = FALSE, status="primary", title = "QTL profile",
                      column(12,
                             box(
-                              title = "Required inputs", width = 5, background = "light-blue",
+                              title = "Required inputs (*)", width = 5, background = "light-blue",
                               "* QTL analysis files or viewpoly object or example dataset (check `Input data` tab)"
                             )
                      ), 
@@ -86,7 +86,7 @@ mod_qtl_view_ui <- function(id){
                      box(width = 12, solidHeader = FALSE, collapsible = TRUE,  collapsed = TRUE, status="primary", title = "Effects",
                          column(12,
                                 box(
-                                  title = "Required inputs", width = 5, background = "light-blue",
+                                  title = "Required inputs (*)", width = 5, background = "light-blue",
                                   "* QTL analysis files or viewpoly object or example dataset (check `Input data` tab)", br(),
                                   "* Selection of QTL/s (triangle/s at the bottom of QTL profile graphic)"
                                 )
@@ -119,7 +119,7 @@ mod_qtl_view_ui <- function(id){
                      box(width = 12, solidHeader = FALSE, collapsible = TRUE,  collapsed = TRUE, status="primary", title = "Progeny haplotypes",
                          column(12,
                                 box(
-                                  title = "Required inputs", width = 5, background = "light-blue",
+                                  title = "Required inputs (*)", width = 5, background = "light-blue",
                                   "* QTLpoly analysis files or viewpoly object or example dataset (check `Input data` tab)", br(),
                                   "* Selection of QTL/s (triangle/s at the bottom of QTL profile graphic)"
                                 )
@@ -128,7 +128,7 @@ mod_qtl_view_ui <- function(id){
                                 actionBttn(ns("haplo_update"), style = "jelly", color = "royal",  size = "sm", label = "update available haplotypes", icon = icon("refresh", verify_fa = FALSE)), 
                                 br(), br(),
                                 pickerInput(ns("haplo"),
-                                            label = h6("Select haplotypes"),
+                                            label = h6("Select haplotypes*"),
                                             choices = "Click on `update available haplotype` to update",
                                             selected = "Click on `update available haplotype` to update",
                                             options = pickerOptions(
@@ -140,7 +140,7 @@ mod_qtl_view_ui <- function(id){
                                               dropdownAlignRight = TRUE
                                             ), 
                                             multiple = TRUE), br(),
[…]
