+++ R/mod_map_view.R
@@ -32,7 +32,7 @@ mod_map_view_ui <- function(id){
           
           column(6,
                  column(6,
-                        box(width = 12, solidHeader = TRUE,  status="info", title = "Select phenotypes",
+                        box(width = 12, solidHeader = TRUE,  status="info", title = "Select phenotypes*",
                             pickerInput(ns("phenotypes"),
                                         label = h4("Phenotypes:"),
                                         choices = "This will be updated",
@@ -46,7 +46,7 @@ mod_map_view_ui <- function(id){
                         ), br(),
                  ),
                  column(6,
-                        box(width = 12, solidHeader = TRUE, status="info", title = "Select linkage group",
+                        box(width = 12, solidHeader = TRUE, status="info", title = "Select linkage group*",
                             selectInput(inputId = ns("group"), label = p("Linkage group:"), choices = 1:15, selected = 1),
                             checkboxInput(ns("op"), label = "Show SNP names", value = TRUE)
                         ), br(),
@@ -61,7 +61,7 @@ mod_map_view_ui <- function(id){
         box(width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = TRUE, status="primary", title = "QTL profile",
             column(12,
                    box(
-                     title = "Required inputs", width = 5, background = "light-blue",
+                     title = "Required inputs (*)", width = 5, background = "light-blue",
                      "* QTL analysis files or viewpoly object or example dataset (check `Input data` tab)", 
                    )
             ), 
@@ -88,7 +88,7 @@ mod_map_view_ui <- function(id){
         box(width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = FALSE, status="primary", title = "Map",
             column(12,
                    box(
-                     title = "Required inputs", width = 5, background = "light-blue",
+                     title = "Required inputs (*)", width = 5, background = "light-blue",
                      "* Linkage map files or viewpoly object or example dataset (check `Input data` tab)", 
                    )
             ), 
@@ -119,7 +119,7 @@ mod_map_view_ui <- function(id){
         box(width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = TRUE, status="primary", title = "Map summary",
             column(12,
                    box(
-                     title = "Required inputs", width = 5, background = "light-blue",
+                     title = "Required inputs (*)", width = 5, background = "light-blue",
                      "* Linkage map files or viewpoly object or example dataset (check `Input data` tab)", 
                    )
             ), 
