### 2022-05-24 f6a9239
improve required inputs #34
