+++ R/mod_qtl_view.R
@@ -13,13 +13,15 @@
 mod_qtl_view_ui <- function(id){
   ns <- NS(id)
   tagList(
+    useShinyjs(),
+    extendShinyjs(text = jscode, functions = "collapse"),
     fluidPage(
       verticalLayout(
         fluidRow(
           column(width = 12,
                  div(style = "position:absolute;right:1em;", 
                      div(
-                       actionButton(ns("goUploads"), "Go to Input data",icon("arrow-circle-left", verify_fa = FALSE), class = "btn btn-primary"),
+                       actionButton(ns("goUploads"), "Go to Input data", icon("arrow-circle-left", verify_fa = FALSE), class = "btn btn-primary"),
                        actionButton(ns("goGenes"), label = div("Go to Genome", icon("arrow-circle-right", verify_fa = FALSE)), class = "btn btn-primary"))
                  )
           ),
@@ -27,7 +29,7 @@ mod_qtl_view_ui <- function(id){
           column(6,
                  column(12,
                         box(
-                          width = 3, background = "light-blue",
+                          background = "light-blue",
                           "Required inputs (*)", br(),
                         )
                  ),
@@ -61,38 +63,38 @@ mod_qtl_view_ui <- function(id){
                  )
           ),
           column(12,
-                 box(width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = FALSE, status="primary", title = "QTL profile",
+                 box(id = ns("box_profile"), width = 12, solidHeader = TRUE, collapsible = TRUE,  collapsed = FALSE, status="primary", title = actionLink(inputId = ns("profileID"), label = "QTL profile"),
                      column(12,
                             box(
-                              width = 5, background = "light-blue",
+                              background = "light-blue",
                               "* QTL analysis files or viewpoly object or example dataset (check `Input data` tab)"
                             )
                      ), 
                      column(12,
-                            column(1,
+                            column(3,
                                    downloadBttn(ns('bn_download'), style = "gradient", color = "royal")
                             ),
                             column(3,
                                    radioButtons(ns("fformat"), "File type", choices=c("png","tiff","jpeg","pdf"), selected = "png", inline = T)
                             ),                     
                             column(2,
-                                   numericInput(ns("width_profile"), "Width (mm)", value = 180, width = '40%'),
+                                   numericInput(ns("width_profile"), "Width (mm)", value = 180),
                             ),
                             column(2,
-                                   numericInput(ns("height_profile"), "Height (mm)", value = 120, width = '40%'),
+                                   numericInput(ns("height_profile"), "Height (mm)", value = 120),
                             ),
                             column(2,
-                                   numericInput(ns("dpi_profile"), "DPI", value = 300, width = '30%')
+                                   numericInput(ns("dpi_profile"), "DPI", value = 300)
                             )), br(), 
                      column(12,
                             hr(),
                             plotOutput(ns("plot_qtl"), 
                                        click=ns("plot_click"), brush = ns("plot_brush"))
                      ),
-                     box(width = 12, solidHeader = FALSE, collapsible = TRUE,  collapsed = TRUE, status="primary", title = "Effects",
+                     box(id = ns("box_effects"), width = 12, solidHeader = FALSE, collapsible = TRUE,  collapsed = TRUE, status="primary", title = actionLink(inputId = ns("effectsID"), label = "Effects"),
                          column(12,
                                 box(
-                                  width = 5, background = "light-blue",
+                                  background = "light-blue",
                                   "* QTL analysis files or viewpoly object or example dataset (check `Input data` tab)", br(),
                                   "* Selection of QTL/s (triangle/s at the bottom of QTL profile graphic)"
                                 )
[…]
