+++ NAMESPACE
@@ -43,7 +43,9 @@ importFrom(reshape2,melt)
 importFrom(shiny,NS)
 importFrom(shiny,shinyApp)
 importFrom(shiny,tagList)
+importFrom(shinyjs,extendShinyjs)
 importFrom(shinyjs,inlineCSS)
+importFrom(shinyjs,useShinyjs)
 importFrom(stats,chisq.test)
 importFrom(stats,dhyper)
 importFrom(tidyr,pivot_longer)

[…]
