+++ R/mod_genes_view.R
@@ -313,34 +313,33 @@ mod_genes_view_server <- function(input, output, session,
   
   # Plot QTL profile
   output$plot_qtl <- renderPlotly({
-    if(!is.null(loadQTL())){
-      idx <- which(unique(loadQTL()$profile$pheno) %in% input$phenotypes)
-      pl <- plot_profile(profile = loadQTL()$profile, qtl_info = loadQTL()$qtl_info, selected_mks = loadQTL()$selected_mks,
-                         pheno.col = idx,
-                         lgs.id = as.numeric(input$group),
-                         range.min = input$range[1],
-                         range.max = input$range[2], 
-                         by_range=T, 
-                         software = loadQTL()$software)
-      ggplotly(source = "qtl_profile", pl, tooltip=c("Trait","Position (cM)")) %>% 
-        layout(legend = list(orientation = 'h', y = -0.3), 
-               modebar = list(
-                 remove = c("toImage", 
-                            "hovercompare", 
-                            "hoverCompareCartesian")),
-               clickmode ="none",
-               dragmode = FALSE)
-    } else 
-      stop(safeError("Upload the QTL information in upload session to access this feature."))
+    validate(
+      need(!is.null(loadQTL()), "Upload the QTL information in upload session to access this feature.")
+    )
+    idx <- which(unique(loadQTL()$profile$pheno) %in% input$phenotypes)
+    pl <- plot_profile(profile = loadQTL()$profile, qtl_info = loadQTL()$qtl_info, selected_mks = loadQTL()$selected_mks,
+                       pheno.col = idx,
+                       lgs.id = as.numeric(input$group),
+                       range.min = input$range[1],
+                       range.max = input$range[2], 
+                       by_range=T, 
+                       software = loadQTL()$software)
+    ggplotly(source = "qtl_profile", pl, tooltip=c("Trait","Position (cM)")) %>% 
+      layout(legend = list(orientation = 'h', y = -0.3), 
+             modebar = list(
+               remove = c("toImage", 
+                          "hovercompare", 
+                          "hoverCompareCartesian")),
+             clickmode ="none",
+             dragmode = FALSE)
   })
   
   # cM x Mb
   output$plot_pos <- renderPlotly({
-    if(!is.null(loadMap()))
-      if(loadMap()$software == "polymapR") stop(safeError("Feature not implemented for software polymapR."))
-    
-    if(is.null(loadMap()$ph.p1)) stop(safeError("Upload map information in the upload session to access this feature."))
-    
+    validate(
+      need(!(!is.null(loadMap()) & loadMap()$software == "polymapR"), "Feature not implemented for software polymapR."),
+      need(!is.null(loadMap()$ph.p1), "Upload map information in the upload session to access this feature.")
+    )
     p <- plot_cm_mb(loadMap(), input$group, input$range[1], input$range[2])
     
     max_updated = reactive({
@@ -402,32 +401,32 @@ mod_genes_view_server <- function(input, output, session,
       } else path.wig <- NULL
     } else path.wig <- NULL
     
-    if(is.null(loadJBrowse_fasta()) & !is.null(loadExample())){
-      path.fa <- loadExample()$fasta
-      path.gff <- loadExample()$gff3
-      
-      ext.list <- strsplit(c(loadExample()$fasta,loadExample()$gff3), "[.]")
-      
-      ext <- sapply(ext.list, function(x) {
-        if(x[length(x)] == "gz") paste0(x[length(x)-1], ".",x[length(x)])
[…]
