+++ R/mod_qtl_view.R
@@ -280,20 +280,22 @@ mod_qtl_view_server <- function(input, output, session,
   })
   
   qtl.data <- reactive({
-    if(!is.null(loadQTL())){
-      idx <- which(unique(loadQTL()$profile$pheno) %in% input$phenotypes)
-      
-      withProgress(message = 'Working:', value = 0, {
-        incProgress(0.3, detail = paste("building graphic..."))
-        pl <- plot_profile(profile = loadQTL()$profile, 
-                           qtl_info = loadQTL()$qtl_info, 
-                           selected_mks = loadQTL()$selected_mks,
-                           pheno.col = idx,
-                           lgs.id = as.numeric(input$group), 
-                           by_range=F, plot = F)
-      })
-    } else
-      stop(safeError("Upload the QTL information in upload session to access this feature."))
+    validate(
+      need(length(input$phenotypes) != 0, "Select at least one phenotype"),
+      need(length(input$group) != 0, "Select at least one linkage group"),
+      need(!is.null(loadQTL()), "Upload the QTL information in upload session to access this feature.")
+    )
+    idx <- which(unique(loadQTL()$profile$pheno) %in% input$phenotypes)
+    
+    withProgress(message = 'Working:', value = 0, {
+      incProgress(0.3, detail = paste("building graphic..."))
+      pl <- plot_profile(profile = loadQTL()$profile, 
+                         qtl_info = loadQTL()$qtl_info, 
+                         selected_mks = loadQTL()$selected_mks,
+                         pheno.col = idx,
+                         lgs.id = as.numeric(input$group), 
+                         by_range=F, plot = F)
+    })
   })
   
   output$plot_qtl <- renderPlot({
@@ -304,27 +306,26 @@ mod_qtl_view_server <- function(input, output, session,
   })
   
   effects.data <- reactive({
-    if(!is.null(loadQTL())){
-      if(!is.null(input$plot_brush)){
-        df <- try(brushedPoints(qtl.data()[[2]], input$plot_brush, xvar = "x", yvar = "y.dat"))
-        if(inherits(df, "try-error")) 
-          stop(safeError("Select at least one triangle on the bottom of the QTL profile graphic. The triangles refer to QTL peaks detected. You can click and brush your cursor to select more than one."))
-      } else {
-        stop(safeError("Select at least one triangle on the bottom of the QTL profile graphic. The triangles refer to QTL peaks detected. You can click and brush your cursor to select more than one."))
-      }
-      withProgress(message = 'Working:', value = 0, {
-        incProgress(0.5, detail = paste("Getting data..."))
-        data <- data_effects(qtl_info = loadQTL()$qtl_info, 
-                             effects = loadQTL()$effects,
-                             pheno.col = as.character(df$Trait), 
-                             lgs = df$LG, 
-                             position = df$`Position (cM)`,
-                             groups = as.numeric(input$group),
-                             software = loadQTL()$software,
-                             design = input$effects_design)
-      })
-    } else 
-      stop(safeError("Upload the QTL information in upload session to access this feature."))
+    validate(
+      need(!is.null(loadQTL()), "Upload the QTL information in upload session to access this feature."),
+      need(!is.null(input$plot_brush), "Select at least one triangle on the bottom of the QTL profile graphic. The triangles refer to QTL peaks detected. You can click and brush your cursor to select more than one.")
+    )
+    df <- try(brushedPoints(qtl.data()[[2]], input$plot_brush, xvar = "x", yvar = "y.dat"))
+    
+    validate(
[…]

[…]
