### 2022-05-25 mmollina/viewpoly#34 @Cristianetaniguti
Hello @raivivek , thanks for all your comments and suggestions! I made some modifications according to each topic, and here are my comments:

* With the example dataset, the QTL tab is not working properly. For the "Effects", "Breeding values", and "QTL summary" modules I see an error: Select at least one triangle on the bottom of the QTL profile graphic. The triangles refer to QTL peaks detected. You can click and brush your cursor to select more than one.

The information in “Effects”, “Breeding values” and “QTL summary” are details about the QTL selected in the “QTL profile” box. Therefore, it is not possible to display something before this selection and the presence of a QTL peak in the dataset. The triangles to be selected are on the bottom of the QTL profile if a significative QTL is present in the phenotype and chromosome selected.

* It'd be nice to add a prompt to Exit button before really quitting. I was confused at one point if it'll to return to home / reset session and not quit the entire app. It will also be helpful to move the button to header since it controls the entire app not just a tab.

I replaced the exit button by a reset button in the Input data tab and removed from the others. Indeed, the exit button was redundant once the user can simple close the browser.

Commit: 7ab3ff5f3536ddc651b73df0d9927d6bc7df1c9e

The other buttons, that takes from one tab to another may also be redundant too because clicking in the main menu has the same effect, but we wanted to create a sense of continuity between the tabs: first, explore and identify an interesting QTL; second, go relate the QTL location with genome annotation; and third, check linkage map and marker dosage in that region.

* Increase the DPI of downloaded figure. They are large and as such when reduced to a comfortable size, the text is hard to read.

I added new input buttons to every image to make the width, height and dpi user defined.
Commit: dca97af7338d11e2f74e72562cbdac2464f19cc8

* Next/Previous should clarify that it will go to the new tab and not to the modules within the tab. Could say “Go to QTL tab” or “Go back to Input data” etc

Commit: 7ab3ff5f3536ddc651b73df0d9927d6bc7df1c9e

* In each tab, it will be helpful to indicate required input boxes, if any using asterisks

Commits: 6ed54edc3eb544d70f1d18df252b4f473a2f1c92 ; 77915e82333f7b6ba2726b09d6f6332a991f4fcd ; f6a9239fc7b120459cafcdd78b5aa81e3ac465b3

* Update collapse/expand behavior of module tab to also work with the module header text and not just the +/- sign; it should make the app a lot more functional in my opinion

Commit: 90c61448c1ab6023b11b655c323c85fd044c8a0b

* It is also not clear which modules are expanded or collapsed when I switch/open a tab. For example, in the Map tab, "QTL profile" is collapsed but second module is expanded.

We used this resource to highlight from where the user should start (in the case of the QTL tab) or the unique graphics (in Genome and Map tabs) once QTL profile was only included again on those to help guide the linkage map region selection

* Improve ShinyApp exception handling; should you encounter an error, capture the exception and show a message/prompt on front-end that something happened and one should refresh/restart; along with the error message if possible so that users can report issues to you

Commit: 39cc51a41a8e52913bc7347f4b56aa21957b18d5
Now, it was implemented the `validate` and `need` functions from shiny which return a better message for the user. I will keep tracking errors and creating exceptions for them.

* It'd be nice to have additional input validation. For example, if someone clicks submit VIEWPoly without actually adding data, the instance should not crash. Would recommend checking for file existence in all places where you process data.

Commit: 39cc51a41a8e52913bc7347f4b56aa21957b18d5
