### 2022-05-25 mmollina/viewpoly#34 @raivivek
Hi @Cristianetaniguti — a quick question re: your comment below

> The information in “Effects”, “Breeding values” and “QTL summary” are details about the QTL selected in the “QTL profile” box. Therefore, it is not possible to display something before this selection and the presence of a QTL peak in the dataset. The triangles to be selected are on the bottom of the QTL profile if a significative QTL is present in the phenotype and chromosome selected.

Does the example data contain any significant QTLs? If not, would it be possible to use an example data that showcases most (if not all) of the features in your app?
