+++ R/utils_helpers.R
@@ -725,3 +725,6 @@ shinyjs.collapse = function(boxid) {
 $('#' + boxid).closest('.box').find('[data-widget=collapse]').click();
 }
 "
+
+# Global variables to avoid NOTE
+globalVariables("js")
