### 2022-05-25 mmollina/viewpoly#34 @Cristianetaniguti
Hi @raivivek , the example has some significant QTL. I changed the default selected group to group 5 because it presents one of them.
