+++ tests/testthat/test-shinyTest.R
@@ -4,5 +4,5 @@ test_that("shinytest", {
   # Don't run these tests on the CRAN build servers
   skip_on_cran()
 
-  expect_pass(testApp(appDir = system.file(package = "viewpoly"), testnames = c("github_actions_tests"),compareImages = FALSE))
+  expect_pass(testApp(appDir = gsub("inst", "", system.file(package = "viewpoly")), testnames = c("github_actions_tests"),compareImages = FALSE))
 })
