+++ inst/tests/shinytest/github_actions_tests.R
@@ -25,5 +25,3 @@ app$snapshot(list(output = "genes_view_ui_1-plot_qtl"))
 app$setInputs(viewpoly = "map")
 app$snapshot(list(output = "map_view_ui_1-plot_qtl"))
 app$snapshot(list(output = "map_view_ui_1-parents_haplo"))
-app$snapshot(list(output = "map_view_ui_1-summary"))
-app$snapshot(list(output = "map_view_ui_1-map_summary"))
