### 2022-05-31 #4242 @Cristianetaniguti
Hello, I also finished the modifications according to @raivivek  suggestions. I am about to submit this new version to CRAN. Let me know if you have more suggestions. Thanks!
