### 2022-06-01 mmollina/viewpoly#24 @cpalmer718
Apologies for the long delay, it's been a strange few months. Image generation seems to be working for me now,
