### 2022-06-01 mmollina/viewpoly#35
Resilience to unexpected interactions
Various interactions I can have with the interface (to be clear, deliberately silly interactions, like just clicking some text elements) cause a hard crash, or in the case of the hosted version a "disconnected from server" error. An example:

From a clean R installation:
```r
install.packages("devtools")
library(devtools)
devtools::install(".", dependencies = TRUE)
library(viewpoly)
viewpoly::run_app()
## click on the "Input data" tab
## click on the text "Upload linkage map files"
```

I've found a half dozen other interactions that hard crash like this. I understand that the interactions are not intended ones, but it seems like some of those might better be caught and handled in a more graceful way. This is really a matter of user convenience, I guess: having to relaunch everything because I happened to click on the text instead of the "+" is a bit of a bummer.

I'm not really sure from a dev perspective how much of a nightmare it would be to prevent this kind of behavior in general. But for some of the text elements, for example, it seems like clicking on them when no other interactions have occurred might reasonably result in nothing in particular happening, as opposed to a hard crash and a minimal stack trace.
