+++ NAMESPACE
@@ -8,6 +8,7 @@ import(markdown)
 import(shiny)
 import(shinyWidgets)
 import(shinydashboard)
+import(shinyjs)
 import(vroom)
 importFrom(JBrowseR,JBrowseR)
 importFrom(JBrowseR,JBrowseROutput)
