### 2022-06-01 mmollina/viewpoly#35 @Cristianetaniguti
Hello @cpalmer718 ! Thanks for your feedback. I made a silly mistake in this particular error that you mentioned (forgot to import a package). Now, clicking in "Upload linkage map files" or any box title should do the same thing that the + sign does.
I will keep fixing this type of error once I become aware of them.
