### 2022-06-01 #4242 @Cristianetaniguti
Hello @cpalmer718 , thanks again for your comments. And yes, it is difficult to make the app bulletproof, but I will try to fix these types of errors as I become aware of them. I just fixed the one that you pointed out, thanks for that too!
