+++ NAMESPACE
@@ -8,7 +8,6 @@ import(markdown)
 import(shiny)
 import(shinyWidgets)
 import(shinydashboard)
-import(shinyjs)
 import(vroom)
 importFrom(JBrowseR,JBrowseR)
 importFrom(JBrowseR,JBrowseROutput)
@@ -46,6 +45,7 @@ importFrom(shiny,shinyApp)
 importFrom(shiny,tagList)
 importFrom(shinyjs,extendShinyjs)
 importFrom(shinyjs,inlineCSS)
+importFrom(shinyjs,js)
 importFrom(shinyjs,useShinyjs)
 importFrom(stats,chisq.test)
 importFrom(stats,dhyper)
