+++ DESCRIPTION
@@ -60,7 +60,8 @@ Imports:
     markdown,
     stats,
     rlang,
-    pkgload
+    pkgload,
+    vdiffr
 URL: https://github.com/mmollina/viewpoly
 BugReports: https://github.com/mmollina/viewpoly/issues
 Encoding: UTF-8
@@ -68,7 +69,6 @@ LazyData: true
 RoxygenNote: 7.1.2
 Suggests: 
     testthat (>= 3.0.0),
-    vdiffr,
     shinytest
 Config/testthat/edition: 3
 Language: en-US
