### 2022-06-02 #4242 @cpalmer718
@mikldk yes, dev has indicated that https://github.com/mmollina/viewpoly/issues/35 is resolvable, so I'll install the update, flag any remaining crashes I can find, and then I'll be content. I'll check the update later today.
