+++ R/mod_upload.R
@@ -330,7 +330,7 @@ mod_upload_server <- function(input, output, session, parent_session){
   observeEvent(input$viewpolyupID, {
     js$collapse(ns("box_viewpolyup"))
   })
-
+  
   # Check environment
   observe({
     Objs <- Filter(function(x) inherits(get(x), 'viewpoly' ), ls(envir = .GlobalEnv) )
@@ -670,296 +670,299 @@ mod_upload_server <- function(input, output, session, parent_session){
       } else if (values$upload_state_viewpoly == 'reset') {
         return(NULL)
       } else if(values$upload_state_viewpoly == "uploaded"){
-        validate(
-          need(!is.null(input$viewpoly_input) | !is.null(input$viewpoly_env), 
-               "Upload a viewpoly dataset or select one available in your R environment before submit.")
-        )
-        if(!is.null(input$viewpoly_input)){
+        if(!is.null(input$viewpoly_input) | !is.null(input$viewpoly_env)){ 
+          warning("Upload a viewpoly dataset or select one available in your R environment before submit.")
+          viewpoly.obj <- NULL
+        } else if(!is.null(input$viewpoly_input)){
           temp <- load(input$viewpoly_input$datapath)
           viewpoly.obj <- get(temp)
         } else if(!is.null(input$viewpoly_env)) {
           viewpoly.obj = get(input$viewpoly_env)
-        }
+        }  
         return(viewpoly.obj)
       } 
     })
   })
+
+loadMap_custom = reactive({
+  if(!(is.null(input_map()$dosages) & is.null(input_map()$phases) & is.null(input_map()$genetic_map))){
+    req(input_map()$dosages, input_map()$phases, input_map()$genetic_map)
+    withProgress(message = 'Working:', value = 0, {
+      incProgress(0.5, detail = paste("Uploading custom map data..."))
+      prepare_map_custom_files(input_map()$dosages,
+                               input_map()$phases,
+                               input_map()$genetic_map)
+    })
+  } else NULL
+})
+
+loadMap_mappoly =  reactive({
   
-  loadMap_custom = reactive({
-    if(!(is.null(input_map()$dosages) & is.null(input_map()$phases) & is.null(input_map()$genetic_map))){
-      req(input_map()$dosages, input_map()$phases, input_map()$genetic_map)
-      withProgress(message = 'Working:', value = 0, {
-        incProgress(0.5, detail = paste("Uploading custom map data..."))
-        prepare_map_custom_files(input_map()$dosages,
-                                 input_map()$phases,
-                                 input_map()$genetic_map)
-      })
-    } else NULL
-  })
-  
-  loadMap_mappoly =  reactive({
+  if(!is.null(input_map()$mappoly_in)){
+    withProgress(message = 'Working:', value = 0, {
+      incProgress(0.3, detail = paste("Uploading MAPpoly data..."))
+      prepare_MAPpoly(input_map()$mappoly_in)
+    })
+  } else NULL
+})
+
+loadMap_polymapR =  reactive({
+  if(!(is.null(input_map()$polymapR.dataset) & 
+       is.null(input_map()$polymapR.map))) {
+    req(input_map()$polymapR.dataset, input_map()$polymapR.map)
+    withProgress(message = 'Working:', value = 0, {
+      incProgress(0.1, detail = paste("Uploading polymapR data..."))
+      prepare_polymapR(input_map()$polymapR.dataset, input_map()$polymapR.map, 
[…]
