### 2022-06-02 mmollina/viewpoly#35 @Cristianetaniguti
Hi @cpalmer718 ,

(1) This was the only that I couldn't reproduce here. Does it display any other error message in the console? Anyway, the download will be disabled if the object is empty (see below).

(2) I changed the error with a warning. Now, if the user performs these steps the app doesn't crash, but a message is displayed in the console and in the graphics box in the next tabs.

(3) and (4) I thought that disabling the button in these situations was a really good idea, thanks! I figure out that it is possible to do that with shinyjs package functions.
