### 2022-06-02 mmollina/viewpoly#35 @cpalmer718
@Cristianetaniguti well, deeply mysteriously, I'm getting variable behavior. The first time I tried to reproduce (1) again with upstream patches installed, I got a well-behaved .RData file. So that was encouraging. I then reloaded the app and tried again. This time, the download prompt took a moment to process, and in my general scattered state I alt-tabbed to a different window and switched back a moment later. The download prompt appeared, but for some reason asked if I wanted to download a file called `qwe_download.html`. I downloaded it; it's attached, if you're curious; it also came along with some sort of bizarrely hashed file that I've included in the attachment for posterity. Again, this is just from the same download dataset button. So I shut down the app and tried the alt-tabbing mid-download thing again. This time, I managed to completely freeze both the app and R, and had to kill the process.
[example_strange_files.zip](https://github.com/mmollina/viewpoly/files/8827370/example_strange_files.zip)

I don't know, this sounds like weird GUI nonsense here. I won't block your review based on something like this, it's clearly a bigger issue with shiny or something, and these were admittedly strange behaviors; and it worked the first time. You can check it out if you feel like it, but if you can't reproduce, there's really nothing to be done.

Thanks for checking out the things I mentioned, I appreciate it. Hopefully this will prevent some user difficulties.
