### 2022-06-02 #4242 @cpalmer718
@mikldk I've added the last checkbox. My issues have all been meaningfully addressed; thanks to @Cristianetaniguti for the excellent package and edits, and for following up on all my weird UI attempts.
