### 2022-06-02 mmollina/viewpoly#35 @Cristianetaniguti
@cpalmer718 🤔
It is really mysteriously... I still can't reproduce it here.
Does the update that uses the disable download button avoid this behavior?
