### 2022-06-03 mmollina/viewpoly#35 @cpalmer718
@Cristianetaniguti as far as I can tell, I have all of upstream patched and installed, but it still happens intermittently. I do think the patch helped. It seems like it starts out fine at first, but then something gets messed up in the background and it becomes a problem. I can't reliably trigger it, but it seems like after I close and reopen the app a few times, it gets stuck with the `qwe_download.html` option and stays there. Now even if I reboot R it consistently triggers that download. I'm not sure if I have to reboot to get the correct behavior back or what. It doesn't seem like there's any weird background zombie process or anything. All very mysterious.
