### 2022-06-07 #4242 @Cristianetaniguti
* Package version: [0.2.0](https://github.com/mmollina/viewpoly/releases/tag/v0.2.0)
* DOI: 10.5281/zenodo.6621089
