+++ DESCRIPTION
@@ -7,8 +7,9 @@ Authors@R: c(
                   role = c("aut", "cre"),
                   email = "chtaniguti@tamu.edu"),
            person(given = "Gabriel",
-                  family = "Gesteira",
-                  role = c("aut")),
+                  family = "de Siqueira Gesteira",
+                  role = c("aut"),
+                  email = "gdesiqu@ncsu.edu"),
            person(given = "Jeekin",
                   family = "Lau",
                   role = c("ctb")),
