+++ paper/paper.md
@@ -58,7 +58,7 @@ In recent years, genetic analysis in autopolyploid crops has been gaining attent
 
 # Software implementation
 
-`VIEWpoly` comprises three main modules arranged in an intuitive interface that guides the user through the analysis \autoref{fig:flowchart}. In its current version (V 0.1.1), the package supports `polymapR` [@Bourke2018] and `MAPpoly` [@Mollinari2019; @Mollinari2020] results to display genetic maps. The QTL analysis results can be imported from `QTLpoly` [@Pereira2020], `polyqtlR` [@Bourke2021], and `diaQTL` [@Amadeu2021] packages. Maps and QTL input files can be provided in the RData format or standard text formats (CSV or TSV). `VIEWpoly` is also compatible with genome-related files supported by `JBrowseR`, including FASTA, BED, GFF3, BAM, CRAM, VCF, Wiggle, and bigWig files [@Hershberg2021]. Once the data is uploaded, the user can export the information in a single compressed file to expedite the upload process in further analysis.
+`VIEWpoly` comprises three main modules arranged in an intuitive interface that guides the user through the analysis \autoref{fig:flowchart}. In its current version (V 0.2.0), the package supports `polymapR` [@Bourke2018] and `MAPpoly` [@Mollinari2019; @Mollinari2020] results to display genetic maps. The QTL analysis results can be imported from `QTLpoly` [@Pereira2020], `polyqtlR` [@Bourke2021], and `diaQTL` [@Amadeu2021] packages. Maps and QTL input files can be provided in the RData format or standard text formats (CSV or TSV). `VIEWpoly` is also compatible with genome-related files supported by `JBrowseR`, including FASTA, BED, GFF3, BAM, CRAM, VCF, Wiggle, and bigWig files [@Hershberg2021]. Once the data is uploaded, the user can export the information in a single compressed file to expedite the upload process in further analysis.
 
 ## `VIEWqtl`: QTL Browser
 
